### 2024-01-30 #6307 @editorialbot
[PRE REVIEW]: BoARIO: A Python package implementing the ARIO indirect economic cost model
**Submitting author:** @spjuhel (<a href="http://orcid.org/0000-0001-8801-3890">Samuel Juhel</a>)
**Repository:** https://github.com/spjuhel/BoARIO
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.3
