### 2024-01-31 6088835
missing doi + no double brackets
