+++ paper.bib
@@ -18,18 +18,6 @@
   URL = {http://oe.cd/icio},
 }
 
-@techreport{thissen-euregio-2018,
-  Type = {Working {Paper}},
-  Title = {{EUREGIO}: {The} construction of a global {IO} {DATABASE} with regional detail for {Europe} for 2000-2010},
-  copyright = {http://www.econstor.eu/dspace/Nutzungsbedingungen},
-  ShortTitle = {{EUREGIO}},
-  Abstract = {This paper introduces the EUREGIO database: the first time-series (annual, 2000-2010) of global IO tables with regional detail for the entire large trading bloc of the European Union. The construction of this database, which allows for regional analysis at the level of so-called NUTS2 regions, is presented in detail for its methodology and applications. The tables merge data from WIOD (the 2013 release) with, regional economic accounts, and interregional trade estimates developed by PBL Netherlands Environmental Assessment Agency, complemented with survey-based regional input-output data for a limited number of countries. All used data are survey data and only non-behavioral assumptions have been made to estimate the EUREGIO dataset. These two general rules of data construction allow empirical analyses focused on impacts of changes in behavior (of economies, firms, policies) without endogenously having this behavior embedded already by construction. The tables are publicly available to the research community, from the Dutch government open data website. In this project, both a regional trade database, and production and consumption data of different actors in different regions, were used. The focus is thus intentionally on the regionalization of both trade and the regional use and supply of products by different economic actors. The paper is organized around the successive steps of the data construction: (1) adjustment of WIOD, (2) regional information, (3) Construction of tables, (4) conclusions on the usefulness of this type of regional IO tables with an overview of current applications of the EUREGIO database.},
-  Language = {eng},
-  Number = {TI 2018-084/VI},
-  Institution = {Tinbergen Institute Discussion Paper},
-  Author = {Thissen, Mark and Lankhuizen, Maureen and van Oort, Frank G. and Los, Bart and Diodato, Dario},
-  Year = {2018},
-}
 
 @article{lenzen-2012-mappin-struc,
   Title = {Mapping the Structure of the World Economy},
@@ -380,4 +368,19 @@ https://github.com/IndEcol/pymrio
   Pages = {2520-2522},
   Volume = {28},
   Year = {2012},
+}
+
+@article{thissen-2018-eureg,
+  author =       {Mark Thissen and Maureen Lankhuizen and Frank van Oort and
+                  Bart Los and Dario Diodato},
+  title =        {Euregio: the Construction of a Global Io Database With
+                  Regional Detail for Europe for 2000-2010},
+  journal =      {SSRN Electronic Journal},
+  volume =       {nil},
+  number =       {nil},
+  pages =        {nil},
+  year =         2018,
+  doi =          {10.2139/ssrn.3285818},
+  url =          {http://dx.doi.org/10.2139/ssrn.3285818},
+  DATE_ADDED =   {Wed Jan 31 10:01:29 2024},
 }
