### 2024-01-31 #6307 @spjuhel
:wave: @oliviaguest, and thank you for your time, I'm excited to submit to JOSS !

I made the modifications. Please tell me if you require anything else !
