+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "boario"
-version = "0.5.4"
+version = "0.5.5"
 description = "BoARIO : The Adaptative Regional Input Output model in python."
 authors = ["Samuel Juhel <pro@sjuhel.org>"]
 license = "GNU General Public License v3 or later (GPLv3+)"
