+++ pyproject.toml
@@ -12,7 +12,6 @@ coloredlogs = "^15.0.1"
 country-converter = "^1.0.0"
 dask = ">=2023"
 numpy = "<1.24"
-numpyencoder = "^0.3.0"
 odfpy = "^1.4.1"
 pandas = ">=1.5"
 progressbar2 = "^4.2.0"
