### 2024-02-24 298a73c
Bumping to master (#76)
* 0.5.2 patch (#64)

* (Extended models) :fix: Attempt at fixing #62

Should fix #62

* Fix final demand only NaN

* fix (simulation.py) : typo for final demand record

* bump dep versions

* Fix (event.py) : Household rebuilding NaN corrected (#65)

* version bump 0.5.3

* 🚑 (event) : Working fix for household rebuilding demand

* 💄 Black

* update website

* 📦 🚑 Fixed a problem with multi-events + pandas version (#66)

Rebuilding production was incorrectly distributed in multiple events simulation,
leading to more capital being rebuilt than what was possible. This is 
[…]
