+++ paper.bib
@@ -1,0 +1,386 @@
+@article{stadler18-exiob,
+  author = {Konstantin Stadler and Richard Wood and Tatyana Bulavskaya and Carl-Johan S{\"o}dersten and Moana Simas and Sarah Schmidt and Arkaitz Usubiaga and Jos{\'e} Acosta-Fern{\'a}ndez and Jeroen Kuenen and Martin Bruckner and Stefan Giljum and Stephan Lutter and Stefano Merciai and Jannick H. Schmidt and Michaela C. Theurl and Christoph Plutzar and Thomas Kastner and Nina Eisenmenger and Karl-Heinz Erb and Arjan de Koning and Arnold Tukker},
+  title = {Exiobase 3: Developing a Time Series of Detailed Environmentally Extended Multi-Regional Input-Output Tables},
+  journal = {Journal of Industrial Ecology},
+  volume = {22},
+  number = {3},
+  pages = {502-515},
+  year = {2018},
+  doi = {10.1111/jiec.12715},
+  url = {http://dx.doi.org/10.1111/jiec.12715},
+  DATE_ADDED = {Mon Mar 14 17:34:44 2022},
+}
+
+@misc{oecd-2021-oecd-inter,
+  Author = {OECD},
+  Date = {2021},
+  Title = {OECD Inter-Country Input-Output Database},
+  URL = {http://oe.cd/icio},
+}
+
+
+@article{lenzen-2012-mappin-struc,
+  Title = {Mapping the Structure of the World Economy},
+  Author = {Manfred Lenzen and Keiichiro Kanemoto and Daniel Moran and Arne Geschke},
+  DOI = {10.1021/es300171x},
+  JournalTitle = {Environmental Science \& Technology},
+  Number = {15},
+  Pages = {8374-8381},
+  URL = {https://doi.org/10.1021/es300171x},
+  Volume = {46},
+  Year = {2012},
+  date_added = {Thu Jul 8 11:11:11 2021},
+}
+
+@article{coronese-2022-econom-impac,
+  Title = {Economic Impacts of Natural Hazards and Complexity Science: a Critical Review},
+  Author = {Matteo Coronese and Davide Luzzati},
+  DOI = {10.2139/ssrn.4101276},
+  Journal = {SSRN Electronic Journal},
+  Number = {nil},
+  Pages = {nil},
+  URL = {http://dx.doi.org/10.2139/ssrn.4101276},
+  Volume = {nil},
+  Year = {2022},
+  date_added = {Thu Nov 23 17:17:22 2023},
+}
+
+@article{botzen-2019-econom-impac,
+  Title = {The {{Economic Impacts}} of {{Natural Disasters}}: {{A Review}} of {{Models}} and {{Empirical Studies}}},
+  Abstract = { Abstract Economic losses from natural disasters have been increasing in recent decades. This has been attributed mainly to population and economic growth in disaster-prone areas. Future natural disaster losses are expected to increase due to a continued increase in economic exposure and climate change. This highlights the importance of designing policies that can mitigate the impacts of these disasters on the economy and society. A rapidly expanding literature has estimated the direct (e.g., property damage) and indirect (e.g., gross domestic product growth, trade) economic impacts of natural disasters. This article reviews this emerging literature. We synthesize the main theoretical, computational, and empirical methods used, summarize key findings on the economic impacts of natural disasters, and discuss factors that have been found to mitigate disaster impacts. We conclude by identifying lessons for policymakers and outlining an agenda for future research in this field. },
+  Author = {Botzen, W J Wouter and Deschenes, Olivier and Sanders, Mark},
+  DOI = {10.1093/reep/rez004},
+  Date = {2019-08-01},
[…]
