+++ paper.md
@@ -1,0 +1,151 @@
+---
+title: 'BoARIO: A Python package implementing the ARIO indirect economic cost model'
+tags:
+  - Python
+  - economy
+  - indirect impacts
+  - input-output modeling
+authors:
+  - name: Samuel Juhel
+    orcid: 0000-0001-8801-3890
+    equal-contrib: true
+    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+affiliations:
+ - name: CIRED, France
+   index: 1
+ - name: LMD, France
+   index: 2
+date: 13 August 2017
+bibliography: paper.bib
+
+---
+
+# Summary
+
+The impacts of economic shocks (caused by natural or technological disasters for
+instance) often extend far beyond the cost of their local, direct
+consequences, as the economic perturbations they cause propagate along supply
+chains. Understanding the additional impacts and costs stemming from this
+propagation is key to design efficient risk management policies. The interest is rising
+for the evaluation of these "indirect risks" in the context of
+climate change--which leads to an increase in the average risk of weather extremes
+[@lange-2020-projec-expos], and globalized-just-in-time production processes.
+Such evaluations rely on dynamic economic models that represent the interactions
+between multiple regions and sectors. Recent research in the field argues in
+favor of using more Agent-Based oriented model, associated with an increase in
+the complexity of the mechanisms represented [@coronese-2022-econom-impac].
+However, the assumptions and hypotheses underlying these economic mechanisms
+vary a lot, and sometime lack transparency, making it difficult to properly
+interpret and compare results across models, even more so when the code used is
+not published or undocumented.
+
+The Adaptive Regional Input-Output model (or ARIO) is an hybrid input-output /
+agent-based economic model, designed to compute indirect costs consequent to
+economic shocks. Its first version dates back to 2008 and was originally developed to assess the indirect costs of natural disasters
+[@hallegatte-2008-adapt-region]. ARIO is now a well-established and pivotal
+model in its field, has been used in multiple studies, and has seen several
+extensions or adaptations [@wu-2011-region-indir; @ranger-2010-asses-poten;
+@henriet-2012-firm-networ; @hallegatte-2013-model-role;
+@hallegatte-2010-asses-climat; @hallegatte-2008-adapt-region;
+@guan-2020-global-suppl; @jenkins-2013-indir-econom; @koks-2015-integ-direc;
+@wang-2020-econom-footp; @wang-2018-quant-spatial].
+
+In ARIO, the economy is modelled as a set of economic sectors and regions, and
+we call a specific (region,sector) couple an *industry*. Each industry produces
+a unique product which is assumed to be the same for all industries of the same
+sector. Each industry keeps an inventory of inputs it requires for production.
+Each industry answers a total demand consisting of the final demand (from
+households, public spendings and private investments) and of the intermediate
+demand (from other industries). An initial equilibrium state for the economy is
+built based on a multi-regional input-output table. The model can then describe
+how the economic, as depicted, responds to a shock (or multiple ones).
+
+`BoARIO` is an open-source Python package implementing the ARIO model. Its core
+purpose is to help support better accessibility, transparency, replicability and
+comparability in the field of indirect economic impacts modeling.
+
+# Statement of need
+
+The `BoARIO` package allows to easily run simulations with the ARIO model, via
+simple steps:
+- Instantiating a model
+- Defining one or multiple events
+- Creating a simulation instance that will wrap the model and events, allow to
+  run the simulation, and explore the results.
+
+The ARIO model relies on Multi-Regional Input-Output Tables (MRIOTs) to define
+the initial state of the economy. `BoARIO` was designed to be entirely agnostic
+of the MRIOT used, thanks to the `pymrio` package [@stadler2021_Pymrio]. This
[…]
