+++ boario/event.py
@@ -976,11 +976,9 @@ class EventKapitalDestroyed(Event, ABC):
                         f"households impact ({households_impact}) given as scalar, distributing among region following `impact_regional_distrib` ({self.impact_regional_distrib}) to {self.aff_regions}"
                     )
                     logger.debug(f"{rebuilding_demand_idx}")
-                    self.households_impact_df.loc[
-                        self.aff_regions, rebuilding_demand_idx
-                    ] = (
+                    self.households_impact_df.loc[:, rebuilding_demand_idx] = (
                         households_impact * self.impact_regional_distrib
-                    )  # type: ignore
+                    ).to_numpy()  # type: ignore
             self.households_impact_df *= (
                 self.event_monetary_factor / self.model_monetary_factor
             )
