### 2024-02-24 8642d21
update black action to run only on master during pull request
