+++ .github/workflows/black.yml
@@ -1,12 +1,15 @@
 name: Lint
 
-on: [push, pull_request]
-
+on:
+  pull_request:
+    branches:
+      - master
+  
 jobs:
   lint:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
       - uses: psf/black@stable
         with:
           options: "--check --verbose"
