### 2024-02-24 138cadf
Replaced groupby(axis=1) by .T
