+++ boario/event.py
@@ -687,7 +687,8 @@ class Event(ABC):
     def impact_industries_distrib(self, value: pd.Series):
         self._impact_industries_distrib = value
         self._impact_regional_distrib = self._impact_industries_distrib.groupby(
-            "region", observed=False,
+            "region",
+            observed=False,
         ).sum()
 
     def __repr__(self):
