+++ .github/workflows/black.yml
@@ -1,17 +1,0 @@
-name: Lint
-
-on:
-  pull_request:
-    branches:
-      - master
-  
-jobs:
-  lint:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: psf/black@stable
-        with:
-          options: "--check --verbose"
-          src: "./boario"
-          
