### 2024-02-24 81b2486
Removed python 3.8 from test matrix
We state in pyproject.toml that we require >3.9
