### 2024-02-24 2cce3cc
FIxed test not running on windows
