+++ .github/workflows/CI.yml
@@ -12,18 +12,14 @@ jobs:
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v4
-      - uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
       - name: Run image
         uses: abatilo/actions-poetry@v2
         with:
           poetry-version: ${{ matrix.poetry-version }}
-      - uses: actions/cache@v3
-        name: Define a cache for the virtual environment based on the dependencies lock file
+      - uses: actions/setup-python@v5
         with:
-          path: ./.venv
-          key: venv-${{ hashFiles('poetry.lock') }}
+          python-version: ${{ matrix.python-version }}
+          cache: 'poetry'
       - name: Install the project dependencies
         run: poetry install
       - name: Run black
