### 2024-02-24 181f12d
simplified poetry install
