+++ .github/dependabot.yml
@@ -9,3 +9,10 @@ updates:
     directory: "/pyproject.toml" # Location of package manifests
     schedule:
       interval: "weekly"
+      
+  - package-ecosystem: "github-actions"
+    # Workflow files stored in the default location of `.github/workflows`. (You don't need to specify `/.github/workflows` for `directory`. You can use `directory: "/"`.)
+    directory: "/"
+    schedule:
+      # Check for updates to GitHub Actions every weekday
+      interval: "daily"
