### 2024-02-26 db5db1f
Removed indicators.py for the moment (not up to date, needs rework)
