+++ boario/indicators.py
@@ -1,748 +1,0 @@
-# BoARIO : The Adaptative Regional Input Output model in python.
-# Copyright (C) 2022  Samuel Juhel
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-from __future__ import annotations
-
-from typing import Optional, Union
-
-from boario.extended_models import ARIOPsiModel
-from boario.simulation import Simulation
-import json
-import pathlib
-import numpy as np
-import pandas as pd
-import itertools
-from boario.utils import misc
-import dask.dataframe as da
-from boario import logger
-from boario.event import *
-
-__all__ = ["Indicators"]
-
-
-def create_df(data, regions, sectors):
-    df = pd.DataFrame(
-        data,
-        columns=pd.MultiIndex.from_product(
-            [regions, sectors], names=["region", "sector"]
-        ),
-    )
-    df = df.interpolate()
-    df = df.rename_axis("step")
-    return df
-
-
-def df_from_memmap(memmap: np.memmap, indexes: dict) -> pd.DataFrame:
-    a = pd.DataFrame(
-        memmap,
-        columns=pd.MultiIndex.from_product([indexes["regions"], indexes["sectors"]]),
-    )
-    a["step"] = a.index
-    return a
-
-
-def stock_df_from_memmap(
-    memmap: np.memmap, indexes: dict, timesteps: int, timesteps_simulated: int
-) -> pd.DataFrame:
-    a = pd.DataFrame(
-        memmap.reshape(timesteps * indexes["n_sectors"], -1),
-        index=pd.MultiIndex.from_product(
-            [timesteps, indexes["sectors"]], names=["step", "stock of"]
-        ),
-        columns=pd.MultiIndex.from_product([indexes["regions"], indexes["sectors"]]),
-    )
-    a = a.loc[pd.IndexSlice[:timesteps_simulated, :]]
-    return a
-
-
-class Indicators(object):
-    record_files_list = [
-        "final_demand",
-        "intermediate_demand",
-        "final_demand_unmet",
-        "production_capacity",
-        "production_realised",
-        "limiting_inputs",
-        "overproduction",
-        "rebuild_demand",
-        "rebuild_prod",
-        "productive_capital_to_recover",
-    ]
-
-    params_list = ["simulated_params", "simulated_events"]
-
-    def __init__(
-        self,
-        *,
-        has_crashed,
-        params,
-        regions,
-        sectors,
-        productive_capital,
-        prod,
-        prodmax,
-        overprod,
-        final_demand,
-        io_demand,
-        r_demand,
-        r_prod,
-        fd_unmet,
-        limiting_stocks,
-        n_temporal_units_to_sim: int,
-        n_temporal_units_by_step: int,
-        events: list[Event],
-        results_storage: Optional[str | pathlib.Path] = None,
-        psi_param: Optional[float] = None,
-        inventory_restoration_tau: Optional[int | list[int]] = None,
-        stocks=None,
-        stocks_treatment=False,
-        include_crash: bool = False,
-    ) -> None:
-        logger.info("Instantiating indicators")
-        if not include_crash:
-            if has_crashed:
-                raise RuntimeError(
[…]
