### 2024-02-26 5a6accc
📦 Added several cool packages to dev dependencies in poetry
