+++ pyproject.toml
@@ -30,6 +30,11 @@ sphinx-automodapi = "^0.15.0"
 sphinx-copybutton = "^0.5.2"
 pydata-sphinx-theme = "^0.13.3"
 nbsphinx = "^0.9.2"
+pytest-cov = "^4.1.0"
+isort = "^5.13.2"
+flake8 = "^7.0.0"
+flake8-pyproject = "^1.2.3"
+safety = "^3.0.1"
 
 [build-system]
 requires = ["poetry-core"]
@@ -82,6 +87,13 @@ exclude = ["**/node_modules",
     "tests/",
     "dist/"
 ]
+
+[tool.flake8]
+# Recommend matching the black line length (default 88),
+# rather than using the flake8 default of 79:
+max-line-length = 88
+extend-ignore = ['E501','E203']
+
 
 defineConstant = { DEBUG = true }
 stubPath = "boario/stubs"
