### 2024-02-26 262236a
Bumping to v0.5.5
* bumping version

* removed numpyencoder dependency (#75)

* Fixed groupby(axis=1) FutureWarning (except for pymrio dep)

* Removed indicators.py for the time being (not used currently)

* updated poetry.lock

* Presumably fixes #73

* CI/CD update

   - update black action to run only on master during pull request
   - Improved caching of venv

* Badges !

* updated documentation

* Created dependabot.yml
