+++ boario/model_base.py
@@ -15,10 +15,12 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 from __future__ import annotations
+
 import json
 import pathlib
 import typing
 from typing import Optional
+
 import numpy as np
 import numpy.typing as npt
 import pandas as pd
@@ -27,12 +29,11 @@ from pymrio.core.mriosystem import IOSystem
 from boario import logger, warn_once
 from boario.event import (
     Event,
+    EventArbitraryProd,
     EventKapitalDestroyed,
     EventKapitalRebuild,
-    EventArbitraryProd,
 )
 from boario.utils.misc import lexico_reindex
-
 
 __all__ = [
     "ARIOBaseModel",
@@ -157,7 +158,7 @@ class ARIOBaseModel:
 
         logger.debug("Initiating new ARIOBaseModel instance")
         super().__init__()
-        ################ Parameters variables #######################
+        # ############### Parameters variables #######################
         try:
             logger.info(
                 "IO system metadata :\n{}\n{}\n{}\n{}".format(
@@ -264,10 +265,12 @@ class ARIOBaseModel:
                 [] if infinite_inventories_sect is None else infinite_inventories_sect
             )
             self.inventories = [
-                np.inf
-                if inventory_dict[k] in ["inf", "Inf", "Infinity", "infinity"]
-                or k in infinite_inventories_sect
-                else inventory_dict[k]
+                (
+                    np.inf
+                    if inventory_dict[k] in ["inf", "Inf", "Infinity", "infinity"]
+                    or k in infinite_inventories_sect
+                    else inventory_dict[k]
+                )
                 for k in sorted(inventory_dict.keys())
             ]
 
@@ -282,7 +285,7 @@ class ARIOBaseModel:
             self.inv_duration[self.inv_duration <= 1] = 2
         #################################################################
 
-        ######## INITIAL MRIO STATE (in step temporality) ###############
+        # ####### INITIAL MRIO STATE (in step temporality) ###############
         if pym_mrio.Z is None:
             raise ValueError(
                 "Z attribute of given MRIO doesn't exist, this is a problem"
@@ -343,7 +346,7 @@ class ARIOBaseModel:
             )
 
         value_added[value_added < 0] = 0.0
-        self.gdp_df = value_added.groupby("region", axis=1).sum()
+        self.gdp_df = value_added.T.groupby("region").sum().T
         r"""pandas.DataFrame: Dataframe of the total GDP of each region of the model"""
 
         self.VA_0 = value_added.to_numpy().flatten()
@@ -359,7 +362,7 @@ class ARIOBaseModel:
             if isinstance(self.k_stock, pd.DataFrame):
                 self.k_stock = self.k_stock.squeeze().sort_index().to_numpy()
         elif productive_capital_to_VA_dict is None:
-            logger.warning(f"No capital to VA dictionary given, considering 4/1 ratio")
+            logger.warning("No capital to VA dictionary given, considering 4/1 ratio")
             self.kstock_ratio_to_VA = 4
             self.k_stock = self.VA_0 * self.kstock_ratio_to_VA
         else:
@@ -385,7 +388,7 @@ class ARIOBaseModel:
         r"""numpy.ndarray of float: 2-dim square matrix array of size :math:`(n , n \times m)` representing the threshold under which an input is not considered being an input (0.00001)."""
         #################################################################
 
-        ####### SIMULATION VARIABLES ####################################
+        # ###### SIMULATION VARIABLES ####################################
         self.overprod = np.full(
             (self.n_regions * self.n_sectors), self.overprod_base, dtype=np.float64
         )
@@ -414,7 +417,7 @@ class ARIOBaseModel:
[…]
