+++ .github/workflows/CI.yml
@@ -1,0 +1,25 @@
+name: CI
+on: pull_request
+
+jobs:
+  ci:
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.9", "3.10", "3.11"]
+        os: [ubuntu-latest, macos-latest, windows-latest]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - uses: actions/checkout@v4
+      - name: Install poetry
+        run: pipx install poetry
+      - uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+          cache: 'poetry'
+      - name: Install the project dependencies
+        run: poetry install
+      - name: Run black
+        run: poetry run black --check --verbose ./boario
+      - name: Run the automated tests (for example)
+        run: poetry run pytest -v

[…]
