+++ README.rst
@@ -30,6 +30,9 @@ BoARIO
    :target: https://pypi.org/project/boario/
    :alt: PyPI - Python Version
 
+.. image:: https://joss.theoj.org/papers/71386aa01a292ecff8bafe273b077701/status
+   :target: https://joss.theoj.org/papers/71386aa01a292ecff8bafe273b077701
+   :alt: Joss Status
 
 `BoARIO` : The Adaptative Regional Input Output model in python.
 
