### 2024-03-13 0891921
Replace unnecessary memmap by arrays
