### 2024-03-14 6374972
better record handling
Should be improve even further but it lays the basis to select which variables
to record, and how (just as arrays or as memmaps)
