### 2024-03-14 c8d5a8e
Replace unnecessary memmap by arrays (#84)
* Replace unnecessary memmap by arrays

* black

* Fix order not known

* quick fix for tests

* better record handling

Should be improve even further but it lays the basis to select which variables
to record, and how (just as arrays or as memmaps)

---------

Co-authored-by: sjuhel <pro@sjuhel.org>
