+++ .github/workflows/static_html_deploy.yml
@@ -39,4 +39,4 @@ jobs:
           path: 'docs/build/html'
       - name: Deploy to GitHub Pages
         id: deployment
-        uses: actions/deploy-pages@v1
+        uses: actions/deploy-pages@v4
