### 2024-03-14 c2d04fe
version bump and poetry.toml for venv
