### 2024-03-14 9e2c88c
Update to V0.5.6 (#85)
* poetry dev dependencies update

* Replace unnecessary memmap by arrays (#84)

* Replace unnecessary memmap by arrays

* black

* Fix order not known

* quick fix for tests

* better record handling

Should be improve even further but it lays the basis to select which variables
to record, and how (just as arrays or as memmaps)

---------

Co-authored-by: sjuhel <pro@sjuhel.org>

* Bump actions/upload-pages-artifact from 1 to 3 (#83)

Bumps [actions/upload-pages-artifact](https://github.com/actions/upload-pages-artifact) from 1 to 3.
- [Release notes](https://github.com/
[…]
