### 2024-03-15 c5f9551
trying to fix dependencies for conda forge (#86)
Co-authored-by: sjuhel <pro@sjuhel.org>
