### 2024-03-22 #6307 @spjuhel
:wave: Everyone, @crvernon,

I have made some changes to the package since my initial submission, notably in order to publish it on conda-forge, improve the CI/CD pipeline. The latest version is now 0.5.7, should I update the submission?
