### 2024-03-27 #6547 @editorialbot
[REVIEW]: BoARIO: A Python package implementing the ARIO indirect economic cost model
**Submitting author:** @spjuhel (<a href="http://orcid.org/0000-0001-8801-3890">Samuel Juhel</a>)
**Repository:** https://github.com/spjuhel/BoARIO
**Branch with paper.md** (empty if default branch): main
**Version:** v0.5.10
