### 2024-04-08 #6547 @mwt
@spjuhel Hi, I am going through this right now and have identified some issues.

1. I do not think that the paper explains the state of the field. That is, it does not outline what other software, if any, exists to solve this problem. Otherwise, I think the paper is good.
2. The documentation is incomplete. For example, `EventKapitalRebuild.from_scalar_regions_sectors()`, which is used in the example, is undocumented outside of the automatically generated api reference. It seems to be similar to `EventKapitalRebuild.from_series()`, but it is also not a drop-in replacement.
3. Some documentation appears to be wrong or confusing. For example, the description of `impact_sectoral_distrib` says
    > A vector of equal size to the list of sectors affected, stating the share of the impact each industry should receive. Defaults to None.

    but the input given in the example is the string `"gdp"`, which is not a share, and a list of actual shares like `[0.5, 0.5]` results in a KeyError.

For the second point, I think it would be ideal if the quickstart guide contained all of the information that was needed to understand and execute the quickstart example. This means that we should get a rundown of the events api in the quickstart guide with just enough information to do the example. The rest of the docs can go into more details about the options and alternative ways to specify events.

**Smaller points:**

* In the quickstart example, it is not necessary to import pandas for the plots because it is already loaded inside your package.
* I believe that the sector key `manufactoring` is supposed to be `manufacturing`, but I may be wrong.
* Automatically generated docs are not formatted correctly for arguments. It would be good if the function arguments were nicely formatted instead of running together. This "args" section is also not being detected as the same as the "parameters" section, and so the inputs are enumerated twice. There are multiple ways to do this, but the following format has always worked for me:
  ```
  Parameters
  ----------
  n : int
      Sample size.
  k : int
      Number of moments.
  alpha : float
      Significance level.
  ```
