+++ .github/dependabot.yml
@@ -6,7 +6,8 @@
 version: 2
 updates:
   - package-ecosystem: "pip" # See documentation for possible values
-    directory: "/poetry.lock" # Location of package manifests
+    directory: "/" # Location of package manifests
+    target-branch: "develop"
     schedule:
       interval: "weekly"
       
