### 2024-04-09 8ec51c8
Bump pydata-sphinx-theme from 0.13.3 to 0.15.2 (#91)
* Update dependabot.yml

* Bump pydata-sphinx-theme from 0.13.3 to 0.15.2

Bumps [pydata-sphinx-theme](https://github.com/pydata/pydata-sphinx-theme) from 0.13.3 to 0.15.2.
- [Release notes](https://github.com/pydata/pydata-sphinx-theme/releases)
- [Changelog](https://github.com/pydata/pydata-sphinx-theme/blob/main/RELEASE.md)
- [Commits](https://github.com/pydata/pydata-sphinx-theme/compare/v0.13.3...v0.15.2)

---
updated-dependencies:
- dependency-name: pydata-sphinx-theme
  dependency-type: direct:development
  update-type: version-update
[…]
