+++ pyproject.toml
@@ -28,7 +28,7 @@ sphinxcontrib-bibtex = "^2.5.0"
 sphinx-autoapi = "^2.1.0"
 sphinx-automodapi = "^0.15.0"
 sphinx-copybutton = "^0.5.2"
-pydata-sphinx-theme = "^0.13.3"
+pydata-sphinx-theme = ">=0.13.3,<0.16.0"
 nbsphinx = "^0.9.2"
 pytest-cov = "^4.1.0"
 isort = "^5.13.2"
