+++ boario/event.py
@@ -65,13 +65,13 @@ class Event(ABC):
     # Class Attributes
 
     possible_sectors: pd.Index = pd.Index([])
-    r"""List of sectors present in the MRIO used by the model"""
+    r"""List of sectors present in the MRIOT used by the model"""
 
     possible_regions: pd.Index = pd.Index([])
-    r"""List of regions present in the MRIO used by the model"""
+    r"""List of regions present in the MRIOT used by the model"""
 
     possible_final_demand_cat: pd.Index = pd.Index([])
-    r"""List of final demand categories present in the MRIO used by the model"""
+    r"""List of final demand categories present in the MRIOT used by the model"""
 
     temporal_unit_range: int = 0
     r"""Maximum temporal unit simulated"""
@@ -92,7 +92,7 @@ class Event(ABC):
     r"""lexicographic sector indexes"""
 
     model_monetary_factor: int = 1
-    r"""Amount of unitary currency used in the MRIO (e.g. 1000000 if in € millions)"""
+    r"""Amount of unitary currency used in the MRIOT (e.g. 1000000 if in € millions)"""
 
     gva_df: pd.Series = pd.Series([], dtype="float64")
     r"""GVA per (region,sector)"""
@@ -107,7 +107,7 @@ class Event(ABC):
     r"""Normalized final consumption matrix"""
 
     mrio_name: str = ""
-    r"""MRIO identification"""
+    r"""MRIOT identification"""
 
     @abstractmethod
     def __init__(
@@ -206,19 +206,46 @@ class Event(ABC):
         name: Optional[str] = None,
         **kwarg,
     ) -> Event:
-        """Create an event from an impact pd.Series.
+        """Create an event for an impact given as a pd.Series.
 
-        Args:
-            impact (pd.Series): A vector definition of the impact per (region,sector)
-            occurrence (int, optional): The ordinal of occurrence of the event (requires to be > 0). Defaults to 1.
-            duration (int, optional): The duration of the event (entire impact applied during this number of steps). Defaults to 1.
-            name (Optional[str], optional): A possible name for the event, for convenience. Defaults to None.
+        Parameters
+        ----------
+        impact : pd.Series
+            A pd.Series defining the impact per (region, sector)
+        occurrence : int
+            The ordinal of occurrence of the event (requires to be > 0). Defaults to 1.
+        duration : int
+            The duration of the event (entire impact applied during this number of steps). Defaults to 1.
+        name : Optional[str]
+            A possible name for the event, for convenience. Defaults to None.
+        **kwarg :
+            Keyword arguments
+            Other keyword arguments to pass to the instantiate method (depends on the type of event)
 
-        Raises:
-            ValueError: Raised if impact is empty or contains negative values.
+        Returns
+        -------
+        Event: an Event object or one of its subclass
 
-        Returns:
-            Event: an Event object (or one of its subclass).
+        Raises
+        ------
+        ValueError: Raised if impact is empty of contains negative values.
+
+        Examples
+        --------
+        >>> import pandas as pd
+        >>> import pymrio as pym
+        >>> from boario.simulation import Simulation
+        >>> from boario.extended_model import ARIOPsiModel
+        >>> from boario.event import EventKapitalRecover
+        >>>
+        >>> mriot = pym.load_test()
+        >>> mriot.calc_all()
+        >>>
+        >>> impact_series = pd.Series({('reg1', 'electricity'): 100000.0, ('reg1', 'mining'): 50000.0})
+        >>> model = ARIOPsiModel(mriot)
+        >>> sim = Simulation(model)
[…]
