### 2024-04-10 a3cf0e7
📝 Event doc update [WIP]
