### 2024-04-10 eb54288
🤖 📝 Updated doc building config
Notably fixes some problems with typehints
[WIP] On improving the overal structure of the API
