+++ docs/source/conf.py
@@ -13,14 +13,15 @@
 import os
 import sys
 import re
-sys.path.insert(0, os.path.abspath('../../'))
+
+sys.path.insert(0, os.path.abspath("../../"))
 
 
 # -- Project information -----------------------------------------------------
 
-project = 'BoARIO'
-copyright = '2022, Samuel Juhel'
-author = 'Samuel Juhel'
+project = "BoARIO"
+copyright = "2022, Samuel Juhel"
+author = "Samuel Juhel"
 html_last_updated_fmt = "%b %d, %Y"
 
 # -- General configuration ---------------------------------------------------
@@ -28,34 +29,48 @@ html_last_updated_fmt = "%b %d, %Y"
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.napoleon', 'sphinx.ext.autodoc', 'sphinx.ext.coverage', 'sphinx.ext.intersphinx','sphinx.ext.viewcode', 'sphinx.ext.mathjax','sphinxcontrib.bibtex', 'sphinx.ext.inheritance_diagram',
-#'autoapi.extension',
-#'sphinx_automodapi.automodapi',
- 'sphinx_copybutton', 'nbsphinx']
+extensions = [
+    "numpydoc",
+    "sphinx_copybutton",
+    "sphinx.ext.autodoc",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.coverage",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.viewcode",
+    "sphinx.ext.mathjax",
+    "sphinxcontrib.bibtex",
+    "sphinx.ext.inheritance_diagram",
+    "nbsphinx",
+    'autoapi.sphinx',
+]
 
 numpydoc_show_class_members = False
+numpydoc_show_inherited_class_members = False
+numpydoc_attributes_as_param_list = False
 napoleon_google_docstring = False
 napoleon_use_param = False
-napoleon_use_ivar = True
+napoleon_use_ivar = False
 
 # automatically generate api references
 autosummary_generate = ["boario-api-reference.rst"]
 
 # bibfile
-bibtex_bibfiles = ['refs.bib']
+bibtex_bibfiles = ["refs.bib"]
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["archive.py"]
+exclude_patterns = ["archive.py",
+                    "**.ipynb_checkpoints",
+                    ]
 
 # The master toctree document.
 master_doc = "index"
 
-#copybutton config
+# copybutton config
 copybutton_prompt_text = r">>> |\.\.\. |\$ |In \[\d*\]: | {2,5}\.\.\.: | {5,8}: "
 
 # -- Options for HTML output -------------------------------------------------
@@ -63,69 +78,75 @@ copybutton_prompt_text = r">>> |\.\.\. |\$ |In \[\d*\]: | {2,5}\.\.\.: | {5,8}: 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'pydata_sphinx_theme'
[…]
