+++ docs/build/html/_sources/boario-api-reference.rst.txt
@@ -1,14 +1,23 @@
 .. _api-ref:
 
+.. currentmodule:: boario
+
 #############
 API Reference
 #############
 
-API references for all modules
+This page gives an overview of all public boario objects, functions and
+methods.
 
-.. currentmodule:: boario
+- ``boario.model_base``: Core abstract model definition.
+- ``boario.extended_model``: Actual models definition.
+- ``boario.simulation``: Simulation wrapper.
+- ``boario.event``: Defining different kind of events.
 
 .. toctree::
-   :maxdepth: 4
+   :maxdepth: 2
 
-   autoapi/boario
+   auto/boario.model_base
+   auto/boario.extended_models
+   auto/boario.simulation
+   auto/boario.event
