+++ docs/source/autoapi/boario.simulation.rst
@@ -1,27 +1,0 @@
-=====================
-``boario.simulation``
-=====================
-
-.. automodule:: boario.simulation
-
-   .. contents::
-      :local:
-
-.. currentmodule:: boario.simulation
-
-
-Classes
-=======
-
-- :py:class:`Simulation`:
-  Defines a simulation object with a set of parameters and an IOSystem.
-
-
-.. autoclass:: Simulation
-   :members:
-
-   .. rubric:: Inheritance
-   .. inheritance-diagram:: Simulation
-      :parts: 1
-      
-   .. rubric:: Attributes and Methods
