+++ docs/source/autoapi/boario.indicators.rst
@@ -1,27 +1,0 @@
-=====================
-``boario.indicators``
-=====================
-
-.. automodule:: boario.indicators
-
-   .. contents::
-      :local:
-
-.. currentmodule:: boario.indicators
-
-
-Classes
-=======
-
-- :py:class:`Indicators`:
-  Undocumented.
-
-
-.. autoclass:: Indicators
-   :members:
-
-   .. rubric:: Inheritance
-   .. inheritance-diagram:: Indicators
-      :parts: 1
-      
-   .. rubric:: Attributes and Methods
