+++ pyproject.toml
@@ -7,7 +7,7 @@ license = "GNU General Public License v3 or later (GPLv3+)"
 readme = "README.rst"
 
 [tool.poetry.dependencies]
-python=">=3.9"
+python=">=3.9,<4.0"
 coloredlogs = "^15.0.1"
 country-converter = "^1.0.0"
 dask = ">=2023"
@@ -35,6 +35,11 @@ isort = "^5.13.2"
 flake8 = "^7.0.0"
 flake8-pyproject = "^1.2.3"
 safety = "^3.0.1"
+pydoclint = "^0.4.1"
+ruff = "^0.3.5"
+autoapi = "^2.0.1"
+sphinx-autodoc-typehints = "^2.0.0"
+numpydoc = "^1.7.0"
 
 [build-system]
 requires = ["poetry-core"]
