+++ docs/Makefile
@@ -7,7 +7,6 @@ SPHINXOPTS    ?=
 SPHINXBUILD   ?= sphinx-build
 SOURCEDIR     = source
 BUILDDIR      = build
-DOCS	      = ../new_docs
 
 # Put it first so that "make" without argument is like "make help".
 help:
@@ -19,10 +18,3 @@ help:
 # "make mode" option.  $(O) is meant as a shortcut for $(SPHINXOPTS).
 %: Makefile
 	@$(SPHINXBUILD) -M $@ "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
-
-github:
-	@rm -rf "$(BUILDDIR)"
-	@rm -rf "$(DOCS)/"*
-	@make html
-	@mv -v "$(BUILDDIR)/html/"* "$(DOCS)"
-	@touch "$(DOCS)/.nojekyll"
