### 2024-04-10 8c2e2a9
📝 New autoapi directory
