+++ docs/source/auto/boario.extended_models.rst
@@ -1,0 +1,27 @@
+==========================
+``boario.extended_models``
+==========================
+
+.. automodule:: boario.extended_models
+
+   .. contents::
+      :local:
+
+.. currentmodule:: boario.extended_models
+
+
+Classes
+=======
+
+- :py:class:`ARIOPsiModel`:
+  Undocumented.
+
+
+.. autoclass:: ARIOPsiModel
+   :members:
+
+   .. rubric:: Inheritance
+   .. inheritance-diagram:: ARIOPsiModel
+      :parts: 1
+      
+   .. rubric:: Attributes and Methods
