+++ docs/source/auto/boario.rst
@@ -1,0 +1,21 @@
+==========
+``boario``
+==========
+
+.. automodule:: boario
+
+   .. contents::
+      :local:
+
+
+Submodules
+==========
+
+.. toctree::
+
+   boario.event
+   boario.extended_models
+   boario.model_base
+   boario.simulation
+
+.. currentmodule:: boario
