+++ docs/build/html/_sources/user-guide.rst.txt
@@ -1,12 +1,0 @@
-User guide
-##########
-
-.. toctree::
-   :maxdepth: 1
-
-   boario-what-is
-   boario-installation
-   boario-terminology
-   boario-quickstart
-   boario-events
-   boario-math
