+++ docs/build/html/_sources/boario-math.rst.txt
@@ -1,464 +1,0 @@
-.. _boario-math:
-
-###################################
-Model description
-###################################
-
-Background and overview
-=======================
-	 	 	 	
-Adaptive Regional Input-Output (ARIO) is an I-O model, designed to compute indirect
-costs from exogenous shocks. Its first version dates back to 2008 and has originally
-been developed to assess the indirect costs of natural disasters (Hallegatte 2008).
-In this paper we present BoARIO, a generic python implementation, similar to the one
-described in (Hallegatte 2014) with some additions (e.g. inspired by Guan et al. 2020).
-
-The economy is modelled as a set of economic sectors and a set of regions.
-In the following, we call an industry a specific (sector, region) couple.
-Each economic sector produces its generic product and draws inputs from an inventory.
-Each industry answers to a total demand consisting of a final demand
-(household consumption, public spending and private investments) coming from all
-regions (i.e. both local demand and exports), intermediate demand (inventory resupply).
-An initial equilibrium state of the economy is built based on multi-regional
-input-output tables (MRIO tables). Apart from parameters specific to the MRIOT
-region-sector typology, BoARIO supposedly handles any MRIOT in the same manner.
-
-Multiple kinds of shocks can be implemented: 
-
-* On the production capacity directly (one or multiple industries are arbitrarily
-  forced to produce less)
-* On the productive capital (one or multiple industries arbitrarily 
-  lose some part of their factors of production and are thus forced to both
-  produce less and to build back their capital stock).
-* On the households (households of affected regions lose some part of their goods and
-  seek to get them back).
-
-The model then describes how exogenous shocks propagates across the economy at each
-time step (a day in this study): the imbalance between production and demand is
-resolved by a proportional rationing scheme, which then leads some industries and
-households to not receive the totality of their orders. Industries can buffer
-these unfilled orders with their “inventories” of input and increase or shift
-their resupply demand. Currently, households simply register the loss as
-final consumption not met and do not postpone their demand. During the simulation,
-industries adjust their production and orders based on both demand and their inventories.
-
-Direct economic impact consists in the valuation of the initial exogenous shock,
-while total economic impact includes also indirect costs consequent to the propagation.
-
-Total economic impact can be measured in two ways: 
-
-* Final demand not met, i.e. goods that households couldn’t buy due to rationing.
-* Relative production change, i.e. diminished or increased production of industries relative to their initial production.
-
-
-.. _boario-math-notations:
-
-Notations
-===========
-
-.. include:: notation_table.rstinc
-
-.. note::
-
-  We use sets of indices for industries, sectors and regions to simplify the notation of matrix elements.
[…]
