+++ docs/build/html/_sources/boario-installation.rst.txt
@@ -1,49 +1,0 @@
-###############
-Installation
-###############
-
-.. _installation:
-
-Install via pip
-==================
-
-There is a pip package available, which you can install with:
-
-.. code-block:: console
-
-   pip install boario
-
-
-Otherwise the source-code is available at the GitHub repo: https://github.com/spjuhel/BoARIO
-
-If you encounter any problems while installing the package, please `raise an issue`_
-on the repository, or `contact the developer`_.
-
-.. _raise an issue: https://github.com/spjuhel/BoARIO/issues/new
-
-.. _contact the developer: pro@sjuhel.org
-
-Install from a specific branch
-=================================
-
-Note that you can directly install BoARIO from a branch of its github repository using:
-
-.. code-block:: console
-
-   pip install git+https://github.com/spjuhel/boario.git@branch-name
-
-Requirements
-===============
-
-BoARIO is currently developed using python ``3.9``. It might work with older python versions, but no check were made regarding this.
-
-Installing via pip should install all the required dependencies.
-
-You can also check the ``pyproject.toml`` file in the sources to see all current dependencies.
-
-If you are experiencing difficulties in setting up a working environment feel free to open an issue on the Github repository or to contact the developer.
-
-Platform
-==========
-
-The code is known to run under both Linux and Windows operating systems.
