+++ docs/build/html/_sources/release-note.rst.txt
@@ -1,10 +1,0 @@
-Release notes
-================
-
-0.5.0 (06/2023)
----------------
-
-Features
-*********
-
-* WIP
