+++ docs/build/html/_sources/boario-api-reference.rst.txt
@@ -1,23 +1,0 @@
-.. _api-ref:
-
-.. currentmodule:: boario
-
-#############
-API Reference
-#############
-
-This page gives an overview of all public boario objects, functions and
-methods.
-
-- ``boario.model_base``: Core abstract model definition.
-- ``boario.extended_model``: Actual models definition.
-- ``boario.simulation``: Simulation wrapper.
-- ``boario.event``: Defining different kind of events.
-
-.. toctree::
-   :maxdepth: 2
-
-   auto/boario.model_base
-   auto/boario.extended_models
-   auto/boario.simulation
-   auto/boario.event
