+++ docs/build/html/_sources/notebooks/boario-quickstart.ipynb.txt
@@ -1,119 +1,0 @@
-{
- "cells": [
-  {
-   "attachments": {},
-   "cell_type": "markdown",
-   "id": "ea635dab-2c7f-40e5-b24e-f2ca9ae00de8",
-   "metadata": {},
-   "source": [
-    "## Quickstart example\n",
-    "\n",
-    "In the following example, we use the \"test\" MRIO of the pymrio module,\n",
-    "which doesn't require to load additional data and has a low number of sectors and\n",
-    "regions, allowing fast computation."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 1,
-   "id": "a4f4ad0b-7db2-4b74-b10a-5fd1ce9eb188",
-   "metadata": {},
-   "outputs": [
-    {
-     "data": {
-      "text/plain": [
-       "<Axes: xlabel='step'>"
-      ]
-     },
-     "execution_count": 1,
-     "metadata": {},
-     "output_type": "execute_result"
-    },
-    {
-     "data": {
[…]
