### 2024-04-12 6b9070b
:memo: (event): Doc update
Better description of each type of events

Also some brief descriptions in the main module docstring and model base.
