+++ boario/model_base.py
@@ -14,6 +14,8 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+"""This module defines the core mechanisms of the model."""
+
 from __future__ import annotations
 
 import json
