+++ boario/event.py
@@ -62,7 +62,22 @@ LOW_DEMAND_THRESH = 10
 
 
 class Event(ABC):
-    r"""Base class for events (Abstract)"""
+    r"""An Event object stores all information about a unique shock during simulation
+        such as time of occurrence, duration, type of shock, amount of damages.
+        Computation of recovery or initially requested rebuilding demand is also
+        done in this class.
+
+        .. warning::
+           The Event class is abstract and cannot be instantiated directly. Only its non-abstract subclasses can be instantiated.
+
+        .. note::
+           Events should be constructed using :py:meth:`~Event.from_series()`, :py:meth:`~Event.from_dataframe()`, :py:meth:`~Event.from_scalar_industries()` or from :py:meth:`~Event.from_scalar_regions_sectors()`.
+           Depending on the type of event chosen, these constructors require additional keyword arguments, that are documented for each instantiable Event subclass.
+           For instance, :py:class:`EventKapitalRebuild` additionally requires `rebuild_tau` and `rebuilding_sectors`.
+
+        .. seealso::
+           Tutorial :ref:`boario-events`
+    """
 
     possible_sectors: pd.Index = pd.Index([])
     r"""List of sectors present in the MRIOT used by the model"""
@@ -118,19 +133,6 @@ class Event(ABC):
         occurrence: int = 1,
         duration: int = 1,
     ) -> None:
-        r"""Create an event shocking the model from a dictionary.
-
-        An Event object stores all information about a unique shock during simulation such as
-        time of occurrence, duration, type of shock, amount of damages. Computation
-        of recovery or initially requested rebuilding demand is also done in this
-        class.
-
-        Parameters
-        ----------
-        event : dict
-            A dictionary holding the necessary information to define an event.
-
-        """
         logger.info("Initializing new Event")
         logger.debug("Checking required Class attributes are defined")
 
@@ -214,8 +216,8 @@ class Event(ABC):
         name : Optional[str]
             A possible name for the event, for convenience. Defaults to None.
         **kwarg :
-            Keyword arguments
-            Other keyword arguments to pass to the instantiate method (depends on the type of event)
+            Keyword arguments keyword arguments to pass to the instantiating method
+        (depends on the type of event).
 
         Returns
         -------
@@ -271,7 +273,7 @@ class Event(ABC):
         name: Optional[str] = None,
         **kwarg,
     ) -> Event:
-        """Convenience function for DataFrames. See :meth:`~boario.event.Event.from_series`
+        """Convenience function for DataFrames. See :meth:`~boario.event.Event.from_series`. This constructor only apply ``.squeeze()`` to the given DataFrame.
 
         Parameters
         ----------
@@ -367,8 +369,9 @@ class Event(ABC):
     ) -> Event:
         """Creates an Event from a scalar and a list of industries affected.
 
-        The scalar impact is distributed evenly by default. Otherwise it can be distributed
-        proportionnaly to the GVA of affected industries, or to a custom distribution.
+        The scalar impact is distributed evenly by default. Otherwise it can
+        be distributed proportionnaly to the GVA of affected industries, or to
+        a custom distribution.
 
         Parameters
         ----------
@@ -463,8 +466,11 @@ class Event(ABC):
             A vector of equal size to the list of regions affected, stating the
             share of the impact each industry should receive. Defaults to None.
         impact_sectoral_distrib : Optional[Union[str, npt.ArrayLike]], optional
-            A vector of equal size to the list of sectors affected, stating the
-            share of the impact each industry should receive. Defaults to None.
+            Either:
+
+            * ``\"gdp\"``, the impact is then distributed using the gross value added of each sector as a weight.
[…]
