### 2024-04-12 79d7638
:memo: Update documentation structure, WIP for doc versionning
- Use of autosummary instead of autoapimod for simplicity and per Class page.
Thus update of the module and class templates

- Added a python script to build the doc for multiple versions (untested)
- Added a version switcher to the website
