+++ docs/source/_autosummary/boario.rst
@@ -1,0 +1,47 @@
+﻿boario
+======
+
+.. automodule:: boario
+
+   
+   
+   .. rubric:: Module Attributes
+
+   .. autosummary::
+   
+      DEBUGFORMATTER
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      warn_once
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
+.. rubric:: Modules
+
+.. autosummary::
+   :toctree:
+   :template: custom-module-template.rst
+   :recursive:
+
+   boario.event
+   boario.extended_models
+   boario.model_base
+   boario.simulation
+
