+++ docs/source/_autosummary/boario.model_base.ARIOBaseModel.rst
@@ -1,0 +1,90 @@
+ARIOBaseModel
+=============
+
+.. currentmodule:: boario.model_base
+
+.. autoclass:: ARIOBaseModel
+   :members:
+   :show-inheritance:
+   :inherited-members:
+
+   
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~ARIOBaseModel.__init__
+      ~ARIOBaseModel.calc_inventory_constraints
+      ~ARIOBaseModel.calc_matrix_stock_gap
+      ~ARIOBaseModel.calc_orders
+      ~ARIOBaseModel.calc_overproduction
+      ~ARIOBaseModel.calc_production
+      ~ARIOBaseModel.change_inv_duration
+      ~ARIOBaseModel.distribute_production
+      ~ARIOBaseModel.reset_module
+      ~ARIOBaseModel.update_system_from_events
+      ~ARIOBaseModel.write_index
+   
+   
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~ARIOBaseModel.house_rebuild_demand
+      ~ARIOBaseModel.house_rebuild_demand_tot
+      ~ARIOBaseModel.indus_rebuild_demand
+      ~ARIOBaseModel.indus_rebuild_demand_tot
+      ~ARIOBaseModel.inventory_constraints_act
+      ~ARIOBaseModel.inventory_constraints_opt
+      ~ARIOBaseModel.prod_cap_delta_arbitrary
+      ~ARIOBaseModel.prod_cap_delta_productive_capital
+      ~ARIOBaseModel.prod_cap_delta_tot
+      ~ARIOBaseModel.production_cap
+      ~ARIOBaseModel.production_opt
+      ~ARIOBaseModel.productive_capital_lost
+      ~ARIOBaseModel.tot_rebuild_demand
+      ~ARIOBaseModel.total_demand
+      ~ARIOBaseModel.main_inv_dur
+      ~ARIOBaseModel.regions
+      ~ARIOBaseModel.n_regions
+      ~ARIOBaseModel.sectors
+      ~ARIOBaseModel.n_sectors
+      ~ARIOBaseModel.industries
+      ~ARIOBaseModel.n_industries
+      ~ARIOBaseModel.final_demand_cat
+      ~ARIOBaseModel.n_fd_cat
+      ~ARIOBaseModel.n_temporal_units_by_step
+      ~ARIOBaseModel.iotable_year_to_temporal_unit_factor
+      ~ARIOBaseModel.rebuild_tau
+      ~ARIOBaseModel.overprod_max
+      ~ARIOBaseModel.overprod_tau
+      ~ARIOBaseModel.overprod_base
+      ~ARIOBaseModel.inventories
+      ~ARIOBaseModel.Z_0
+      ~ARIOBaseModel.Z_C
+      ~ARIOBaseModel.Z_distrib
+      ~ARIOBaseModel.Y_C
+      ~ARIOBaseModel.Y_distrib
+      ~ARIOBaseModel.Y_0
+      ~ARIOBaseModel.X_0
+      ~ARIOBaseModel.gdp_df
+      ~ARIOBaseModel.VA_0
+      ~ARIOBaseModel.tech_mat
+      ~ARIOBaseModel.k_stock
+      ~ARIOBaseModel.gdp_share_sector
+      ~ARIOBaseModel.matrix_share_thresh
+      ~ARIOBaseModel.overprod
+      ~ARIOBaseModel.matrix_stock
+      ~ARIOBaseModel.matrix_orders
+      ~ARIOBaseModel.production
+      ~ARIOBaseModel.in_shortage
[…]
