### 2024-04-12 f11ded0
Delete docs/build directory
We're renovating the documentation build
