+++ docs/build/html/_sources/boario-events.rst.txt
@@ -1,293 +1,0 @@
-.. _boario-events:
-
-########################
-How to define Events
-########################
-
-===============
-Introduction
-===============
-
-The BoARIO model is used to study the indirect economic impacts consequent to `local` event [#local]_.
-At the moment, a shock is represented either by:
-
-* A quantity of damages expressed in monetary value and leading to:
-
-  1. A destruction of capital leading to a reduction in production capacity production (See :ref:`boario-math-prod`).
-  2. A possible rebuilding demand, corresponding to destroyed capital or a part of it and addressed to a set of rebuilding sectors.
-
-* Directly by a fraction of production capacity lost for a certain period, which is then recovered exogenously.
-
-To define such an event, we hence need the following information:
-
-1. A quantity of damages (or fraction of production capacity)
-2. The region(s) affected
-3. The sector(s) affected
-4. In the case a rebuilding demand is considered: the sector(s) considered to be rebuilding sectors
-
-In addition, it is also possible to define a duration (corresponding to the time before rebuilding can start)
-as well as time of occurrence (when studying interaction between multiple events), and a recovery duration (:ref:`boario-math-recovery`).
-
-During a simulation BoARIO uses :class:`~boario.event.Event` objects.
-This page details the different aspect of instantiating such objects.
-
-.. [#local] At the moment events are local to the regional unit of the MRIOT used (e.g. a country/world region in the case of EXIOBASE3 MRIO).
-
-Simulation context
-------------------
-
-Creating events requires a simulation context to exist,
-notably to assess the validity of the event (the affected regions, sectors, the occurrence, ...). Hence,
-trying to instantiate an ``Event`` object without already having a ``Simulation`` (and therefore an ``ARIOPsiModel``)
-will raise an error.
-
-In the following, we will assume the following code was run initially:
-
-.. code:: python
-
-        import pymrio
-        import pandas as pd
-
-        from boario.simulation import Simulation
-        from boario.extended_models import ARIOPsiModel
-
-        mrio = pymrio.load_test().calc_all()
-        model = ARIOPsiModel(mrio)
-        sim = Simulation(model, n_temporal_units_to_sim=365)
-
-Here are the regions and sectors defined in this context:
-
-.. code:: pycon
-
-          >>> model.regions
-
-          array(['reg1', 'reg2', 'reg3', 'reg4', 'reg5', 'reg6'], dtype='<U4')
-
-          >>> model.sectors
-
-          array(['construction', 'electricity', 'food', 'manufactoring', 'mining',
-                 'other', 'trade', 'transport'], dtype='<U13')
-
-
-=========================================
-Different types of Event objects
-=========================================
-
-Currently three types of events are implemented and both consider
-a destruction of capital as the impact:
-
-* :class:`~boario.event.EventKapitalRecover` defines events for which the destroyed 
-  capital is recovered (i.e. regained along a specified ``recovery curve`` without an 
-  associated rebuilding demand and rebuilding sectors)
-
-* :class:`~boario.event.EventKapitalRebuild` define events for which the destroyed
-  capital is rebuild (i.e. creates a corresponding rebuilding demand addressed toward
-  a set of rebuilding sectors)
-
-* :class:`~boario.event.EventArbitraryProd` define events for which production capacity 
-  is arbitrarily decrease for a set of industries.
-
-========================================
-Defining Event from an impact vector
-========================================
-
-The more direct way to instantiate an event is to use a :class:`pandas.Series` object
-where the index is the set of affected industries and the values are the impact for each.
-
[…]
