+++ docs/build/html/_sources/boario-api-reference.rst.txt
@@ -1,14 +1,0 @@
-.. _api-ref:
-
-#############
-API Reference
-#############
-
-API references for all modules
-
-.. currentmodule:: boario
-
-.. toctree::
-   :maxdepth: 4
-
-   autoapi/boario
