+++ docs/build/html/_sources/boario-references.rst.txt
@@ -1,37 +1,0 @@
-.. _references:
-
-ARIO Academic work
-==========================
-
-.. image:: https://raw.githubusercontent.com/spjuhel/BoARIO/master/imgs/chronology.svg?sanitize=true
-           :width: 900
-           :alt: ADEME Logo
-
-
-References
-=============
-
-
-:cite:empty:`2008:hallegatte`
-:cite:empty:`2010:ranger`
-:cite:empty:`2011:hallegatte`
-:cite:empty:`2012:henriet`
-:cite:empty:`2013:hallegatte`
-:cite:empty:`2014:wenz`
-:cite:empty:`2015:koks`
-:cite:empty:`2016:koks`
-:cite:empty:`2018:wang`
-:cite:empty:`2018:zhang`
-:cite:empty:`2019:avelino19`
-:cite:empty:`2019:colon`
-:cite:empty:`2019:koks`
-:cite:empty:`2020:guan`
-:cite:empty:`2020:guan`
-:cite:empty:`2020:huang`
-:cite:empty:`2020:mendoza`
-:cite:empty:`2020:shughrue`
-:cite:empty:`2020:wang`
-:cite:empty:`2021:zhang`
-:cite:empty:`2022:huang`
-
-.. bibliography::
