+++ docs/build/html/_sources/index.rst.txt
@@ -1,21 +1,0 @@
-Welcome to BoARIO's documentation!
-##################################
-
-Navigation
-**********
-
-.. toctree::
-   :maxdepth: 1
-
-   user-guide
-   boario-api-reference
-   development
-   release-note
-   boario-references
-
-Indexes
-*******
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
