### 2024-04-12 534abc7
fixed build dir for doc
