+++ .github/workflows/static_html_deploy.yml
@@ -1,10 +1,10 @@
 # Simple workflow for deploying static content to GitHub Pages
-name: Deploy static content to Pages
+name: Build and deploy static content to Pages
 
 on:
   # Runs on pushes targeting the default branch
   push:
-    branches: ["master"]
+    branches: ["develop"]
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
@@ -28,8 +28,23 @@ jobs:
       url: ${{ steps.deployment.outputs.page_url }}
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout
-        uses: actions/checkout@v4
+      - uses: actions/checkout@v4
+      - name: Install poetry
+        run: pipx install poetry
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+          cache: 'poetry'
+      - name: build
+        run: |
+          # ... installation of all tools ...
+          # get all tags, and run a python script
+          # which you find below to build all documents
+          git fetch --tags
+          cd docs
+          poetry run python ./source/build_docs.py
+      - name: Install the project dependencies
+        run: poetry install
       - name: Setup Pages
         uses: actions/configure-pages@v5
       - name: Upload artifact
