+++ .github/workflows/static_html_deploy.yml
@@ -26,6 +26,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+        with:
+          ref: develop
       - name: Install poetry
         run: pipx install poetry
       - uses: actions/setup-python@v5
