+++ .github/workflows/static_html_deploy.yml
@@ -34,6 +34,8 @@ jobs:
         with:
           python-version: "3.10"
           cache: 'poetry'
+      - name: Install the project dependencies
+        run: poetry install
       - name: build
         run: |
           # ... installation of all tools ...
@@ -42,8 +44,6 @@ jobs:
           git fetch --tags
           cd docs
           poetry run python ./source/build_docs.py
-      - name: Install the project dependencies
-        run: poetry install
       - name: Setup Pages
         uses: actions/configure-pages@v5
       - name: Upload artifact
