+++ .github/workflows/static_html_deploy.yml
@@ -4,7 +4,7 @@ name: Build and deploy static content to Pages
 on:
   # Runs on pushes targeting the default branch
   push:
-    branches: ["develop"]
+    branches: ["master"]
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
@@ -30,6 +30,8 @@ jobs:
           ref: develop
       - name: Install poetry
         run: pipx install poetry
+      - name: Install dot
+        run: sudo apt update && sudo apt install -y pandoc graphviz
       - uses: actions/setup-python@v5
         with:
           python-version: "3.10"
