### 2024-04-15 8b06c3a
:memo: Update documentation structure
* Created a Tutorials page instead of the Examples one
* Updated the quickstart example to streamline it
* Separated the differents subsection of the one pager into their own pages for clarity and future improvements of those pages.
* Moved the pages on event definitions
