+++ docs/source/tutorials/model-params.rst
@@ -1,0 +1,97 @@
+.. _model_parameters:
+
+Changing the model parameters
+__________________________________
+
+There are multiple theoretical and implemented versions of the ARIO model, each
+with various parameters. One objective of BoARIO is to offer an extensive,
+modular and adaptable implementation of these versions, in order to allow easy
+access to version comparison, parameters values exploration and modeling
+improvement.
+
+If you are not familiar with the model, it is strongly advised to read the :ref:`boario-math` page of this documentation,
+as well as :cite:`2013:hallegatte`.
+
+Parameters are set when instantiating the model. The following block shows all currently available parameters as well as their default value.
+
+.. code:: python
+
+          model = ARIOPsiModel(
+              pym_mrio=mrio,
+              order_type="alt",
+              alpha_base=1.0,
+              alpha_max=1.25,
+              alpha_tau=365,
+              rebuild_tau=60,
+              main_inv_dur=90,
+              monetary_factor=10**6,
+              temporal_units_by_step=1,
+              iotable_year_to_temporal_unit_factor=365,
+              infinite_inventories_sect=None,
+              inventory_dict=None,
+              productive_capital_vector=None,
+              productive_capital_to_VA_dict=None,
+              psi_param = 0.80,
+              inventory_restoration_tau = 60,
+          )
+
+
+Here a quick description of each parameters. Please refer to both :ref:`the mathematical description<boario-math>` and the :ref:`api-ref` for further details.
+
+* ``order_type`` : Setting it to ``"alt"`` makes the model use the intermediate order mechanism described in :cite:`2020:guan`. Any other value makes the model use the `classic` order mechanism used in :cite:`2013:hallegatte` (see :ref:`alt_orders`)
+
+* ``alpha_base``, ``alpha_max``, ``alpha_tau`` respectively set the base overproduction, the maximum overproduction, and its characteristic time (in `temporal unit`).
+
+* ``rebuild_tau`` sets the default rebuilding or recovering characteristic time for events (this value is overridden if specified directly in the Event object)
+
+* ``inventory_dict`` should be a dictionary of ``sector:duration`` format, where all sector are present and ``duration`` is both the initial and goal duration for this input stock.
+
+* ``main_inv_dur`` sets the default initial/goal inventory duration in `temporal unit` for all sectors if inventory_dict is not given.
+
+* ``infinite_inventories_sect`` should be a list of inputs never constraining production (the stocks for these input will be virtually infinite when considering stock constraints) (overridden by ``inventory_dict``)
+
+* ``monetary_factor`` should be equal to the monetary factor of the MRIO used (most of the time MRIO are in millions €/$, hence the default :math:`10^6`)
+
+* ``temporal_units_by_step`` the number of `temporal units` to simulate every step. Setting it to 5 will divide the computation time by 5, but only one every 5 `temporal units` will actually be simulated. See :ref:`temporal`.
+
+.. _year_to_temporal_unit_factor:
+
+* ``iotable_year_to_temporal_unit_factor`` defines the `temporal unit` assuming the MRIO contains yearly values. Note that this has not been extensively tested and should be used with care.
+
+* ``productive_capital_to_VA_dict`` should be a dictionary of ``sector:ratio`` format, where ratio is an estimate of Capital Stock over Value Added ratio. This is used to estimate the capital stock of each sector. By default the ratio is 4/1 for all sectors.
+
+* ``productive_capital_vector`` can directly set the capital stock for all industries (ie regions*sectors sized). This overrides ``kapital_to_VA_dict``.
+
+* ``psi_param`` and ``inventory_restoration_tau`` : see :ref:`boario-math-dyn`
+
+.. note::
+
+   All arguments except the mrio are keyword arguments (`ie` not positional), meaning you always need to specify <parameter = value>.
[…]
