+++ docs/source/boario-quickstart.rst
@@ -1,292 +1,0 @@
-.. role:: pythoncode(code)
-   :language: python
-
-.. _boario-quickstart:
-
-Quickstart and examples
-==============================
-
-For the following examples, we assume you have followed installation instructions.
-
-If you encounter any problems with one of the examples bellow, please `raise an issue`_
-on the repository, or `contact the developer`_.
-
-.. _raise an issue: https://github.com/spjuhel/BoARIO/issues/new
-
-.. _contact the developer: pro@sjuhel.org
-
-
-Quickstart example
-___________________
-
-.. toctree::
-   :maxdepth: 2
-
-   notebooks/boario-quickstart
-
-ARIO vs :class:`~boario.model_base.ARIOBaseModel` vs :class:`~boario.extended_models.ARIOPsiModel`
-_____________________________________________________________________________________________________________
-
-Currently, two model classes are implemented, :class:`~boario.model_base.ARIOBaseModel` and :class:`~boario.extended_models.ARIOPsiModel`.
-:class:`~boario.model_base.ARIOBaseModel` is essentially a theoretical implementation used to test a "simplistic and essential" version of the model, and should not
-be used directly other than for developing new variants.
-
-:class:`~boario.extended_models.ARIOPsiModel` mostly implements the version presented in :cite:`2013:hallegatte` for the multi-regional case,
-as well as (optionally) the intermediate order mechanism presented in :cite:`2020:guan`.
-
-One mechanism currently not implemented is the `macro effect` on final demand described in :cite:`2008:hallegatte`. This mechanism should be implemented in a future update.
-
-:class:`~pymrio.core.mriosystem.IOSystem` input
-________________________________________________________
-
-The :class:`~pymrio.core.mriosystem.IOSystem` given to instantiate a :class:`~boario.extended_models.ARIOPsiModel` has to
-have intermediate demand matrix ``Z``, final demand matrix ``Y`` and gross production vector ``x`` as attributes, and be balanced.
-
-This should be the case for all MRIO parsed with the ``pymrio`` package.
-Refer to its `documentation <https://pymrio.readthedocs.io/en/latest/>`_ for more details.
-
-.. attention::
-
-   Note that the (region,sector) Multiindexes for the matrices and vector are reordered by BoARIO to be in lexicographic order.
-
-.. _temporal:
-
-Temporal dimension
-______________________
-
-
-The temporal dimension is an important aspect of dynamically modeling indirect economic impacts.
-Historically, ARIO has been used both using weekly and daily steps, but mostly the latter case.
-BoARIO's implementation of ARIO aims at being independent of the ``temporal unit``
-considered, notably to study how this aspect affect results.
-
-This means it is virtually possible to run ARIO on any temporal granularity of your choosing.
-
-For efficiency purpose, this implementation allows to simulate only some ``temporal unit`` and interpolate in between.
-Hence a ``step`` can represent multiple ``temporal units``. Although by default, a ``step`` equals a ``temporal unit`` equals a `day` and
-defaults values of characteristic times and other time related variable are accordingly expressed in number of days, these three terms are conceptually
-different. For this reason we will favor the term ``temporal unit`` to designate the atomic period in the model throughout this documentation.
-
-The number of ``temporal units`` to simulate can be set when instantiating the ``Simulation`` object like so:
-
-.. code:: python
-
-          sim = Simulation(model, n_temporal_units_to_sim=730)
-
-By default, simulation run for 365 `temporal units` which are days by default.
-
-.. _model_parameters:
-
-Changing the model parameters
-__________________________________
-
-
-There are multiple theoretical and implemented versions of the ARIO model, each
-with various parameters. One objective of BoARIO is to offer an extensive,
[…]
