+++ boario/event.py
@@ -63,20 +63,20 @@ LOW_DEMAND_THRESH = 10
 
 class Event(ABC):
     r"""An Event object stores all information about a unique shock during simulation
-        such as time of occurrence, duration, type of shock, amount of damages.
-        Computation of recovery or initially requested rebuilding demand is also
-        done in this class.
+    such as time of occurrence, duration, type of shock, amount of damages.
+    Computation of recovery or initially requested rebuilding demand is also
+    done in this class.
 
-        .. warning::
-           The Event class is abstract and cannot be instantiated directly. Only its non-abstract subclasses can be instantiated.
+    .. warning::
+       The Event class is abstract and cannot be instantiated directly. Only its non-abstract subclasses can be instantiated.
 
-        .. note::
-           Events should be constructed using :py:meth:`~Event.from_series()`, :py:meth:`~Event.from_dataframe()`, :py:meth:`~Event.from_scalar_industries()` or from :py:meth:`~Event.from_scalar_regions_sectors()`.
-           Depending on the type of event chosen, these constructors require additional keyword arguments, that are documented for each instantiable Event subclass.
-           For instance, :py:class:`EventKapitalRebuild` additionally requires `rebuild_tau` and `rebuilding_sectors`.
+    .. note::
+       Events should be constructed using :py:meth:`~Event.from_series()`, :py:meth:`~Event.from_dataframe()`, :py:meth:`~Event.from_scalar_industries()` or from :py:meth:`~Event.from_scalar_regions_sectors()`.
+       Depending on the type of event chosen, these constructors require additional keyword arguments, that are documented for each instantiable Event subclass.
+       For instance, :py:class:`EventKapitalRebuild` additionally requires `rebuild_tau` and `rebuilding_sectors`.
 
-        .. seealso::
-           Tutorial :ref:`boario-events`
+    .. seealso::
+       Tutorial :ref:`boario-events`
     """
 
     possible_sectors: pd.Index = pd.Index([])
@@ -784,19 +784,19 @@ class Event(ABC):
 class EventArbitraryProd(Event):
     r"""An EventArbitraryProd object holds an event with arbitrary impact on production capacity.
 
-        Such events can be used to represent temporary loss of production capacity in a completely exogenous way (e.g., loss of working hours from a heatwave).
+    Such events can be used to represent temporary loss of production capacity in a completely exogenous way (e.g., loss of working hours from a heatwave).
 
-        .. warning::
-           This type of event suffers from a problem with the recovery and does not function properly at the moment.
+    .. warning::
+       This type of event suffers from a problem with the recovery and does not function properly at the moment.
 
-        .. note::
-           For this type of event, the impact value represent the share of production capacity lost of an industry.
+    .. note::
+       For this type of event, the impact value represent the share of production capacity lost of an industry.
 
-        .. note::
-           In addition to the base arguments of an Event, EventArbitraryProd requires a ``recovery_time`` (1 step by default) and a ``recovery_function`` (linear by default).
+    .. note::
+       In addition to the base arguments of an Event, EventArbitraryProd requires a ``recovery_time`` (1 step by default) and a ``recovery_function`` (linear by default).
 
-        .. seealso::
-           Tutorial :ref:`boario-events`
+    .. seealso::
+       Tutorial :ref:`boario-events`
     """
 
     def __init__(
@@ -976,17 +976,17 @@ class EventArbitraryProd(Event):
 class EventKapitalDestroyed(Event, ABC):
     r"""EventKapitalDestroyed is an abstract class to hold events with where some capital (from industries or households) is destroyed. See :py:class:`EventKapitalRecover` and :py:class:`EventKapitalRebuild` for its instantiable classes.
 
-        .. note::
[…]
