+++ pyproject.toml
@@ -36,7 +36,6 @@ flake8 = "^7.0.0"
 flake8-pyproject = "^1.2.3"
 safety = "^3.0.1"
 pydoclint = "^0.4.1"
-ruff = "^0.3.5"
 autoapi = "^2.0.1"
 sphinx-autodoc-typehints = "^2.0.0"
 numpydoc = "^1.7.0"
