+++ pyproject.toml
@@ -25,7 +25,7 @@ pygit2 = "^1.12.1"
 ipykernel = "^6.23.0"
 black = ">=23.3,<25.0"
 sphinxcontrib-bibtex = "^2.5.0"
-sphinx-autoapi = "^2.1.0"
+sphinx-autoapi = ">=2.1,<4.0"
 sphinx-automodapi = "^0.15.0"
 sphinx-copybutton = "^0.5.2"
 pydata-sphinx-theme = ">=0.13.3,<0.16.0"
