### 2024-04-15 #6547 @spjuhel
@mwt Hi, thank you a lot for your remarks,

Indeed, the documentation suffered from several problems.

I did some redesigning, and it should be much better now (Both API and User guide), no huge change for the User guide, but some restructuring. The rework on the API focused on the event module to bring more clarity to the API, and I also switched to a less cluttered and more tree oriented documentation.

In parallel, I should have addressed all your specific points regarding the documentation. See below for more details if you want.

I will push these as a minor patch on the main branch today or tomorrow, alongside an addition in the paper.md regarding the state of the field. Am I correct in thinking this should go into the State of need section?

------------------------

> 1. The documentation is incomplete. For example, EventKapitalRebuild.from_scalar_regions_sectors(), which is used in the example, is undocumented outside of the automatically generated api reference. It seems to be similar to EventKapitalRebuild.from_series(), but it is also not a drop-in replacement.

I think this was due to the function being defined only for the Event main class, the documentation now shows in the subclasses as well, and I have added a lot of information on each subclass specificities notably around the differences in instantiation. Hopefully this is clearer now.

> 2. Some documentation appears to be wrong or confusing. For example, the description of impact_sectoral_distrib says
> > A vector of equal size to the list of sectors affected, stating the share of the impact each industry should receive. Defaults to None.

> but the input given in the example is the string "gdp", which is not a share, and a list of actual shares like [0.5, 0.5] results in a KeyError.

I fixed the documentation. I did not have a KeyError on my side, it is possible this was fixed at some point, but I'm unsure... Could you share a minimal example and the version used?

> 3. In the quickstart example, it is not necessary to import pandas for the plots because it is already loaded inside your package.

Yes indeed! I removed this.

> 4. I believe that the sector key manufactoring is supposed to be manufacturing, but I may be wrong.

Sadly this is a typo from the pymrio package, I will probably suggest a PR on their repo at some point.

> 5. Automatically generated docs are not formatted correctly for arguments. It would be good if the function arguments were nicely formatted instead of running together. This "args" section is also not being detected as the same as the "parameters" section, and so the inputs are enumerated twice.

Yes, I think this came from automated docstring generation I tested at some point and I did not realize it was not in the same format. This should be fixed in the minor patch.
