### 2024-04-16 903cfec
update paper.md and paper.bib
Added a paragraph on other existing implementation
in state of need section.
