### 2024-04-16 631ecc8
update build_doc so that latest doc is in root html
