### 2024-04-16 19cb907
📝 Documentation update
- Added conda-forge install
- updated some urls
- updated python version and OS supported.
- Some minor changes here and there (e.g., mrio -> mriot)
