+++ docs/source/tutorials/mriot-input.rst
@@ -7,8 +7,9 @@ MRIOT Input
 Using a :class:`~pymrio.core.mriosystem.IOSystem` as input
 ________________________________________________________________
 
-The :class:`~pymrio.core.mriosystem.IOSystem` given to instantiate a :class:`~boario.extended_models.ARIOPsiModel` has to
-have intermediate demand matrix ``Z``, final demand matrix ``Y`` and gross production vector ``x`` as attributes, and be balanced.
+The :class:`~pymrio.core.mriosystem.IOSystem` given to instantiate a model requires to
+contain intermediate demand matrix ``Z``, final demand matrix ``Y`` and gross production
+vector ``x`` as attributes, and to be balanced.
 
 This should be the case for all MRIOT parsed with the ``pymrio`` package.
 Refer to its `documentation <https://pymrio.readthedocs.io/en/latest/>`_ for more details.
