### 2024-04-16 e17e664
Improved model lit review section
