+++ docs/source/versions.yaml
@@ -1,3 +1,8 @@
+"latest":
+  tag: '0.5.8'
+  languages:
+    - "en"
+
 "develop":
   tag: 'develop'
   languages:
