### 2024-04-16 34409b0
v0.5.8 (#102)
* Delete test-requirements.txt

This file is no longer used

* Delete pkg-requirements.txt

This file is no longer used

* Clearer badges

Fix breaklines in badge

* Update dependabot.yml

* Update dependabot.yml as in master

* Bump actions/configure-pages from 4 to 5 (#87)

Bumps [actions/configure-pages](https://github.com/actions/configure-pages) from 4 to 5.
- [Release notes](https://github.com/actions/configure-pages/releases)
- [Commits](https://github.com/actions/configure-pages/compare/v4...v5)

---
updated-dependencies:
- dependency-name: actions/configure-pages
  depen
[…]
