+++ paper.bib
@@ -384,3 +384,68 @@ https://github.com/IndEcol/pymrio
   url =          {http://dx.doi.org/10.2139/ssrn.3285818},
   DATE_ADDED =   {Wed Jan 31 10:01:29 2024},
 }
+
+@article{koks-2016-multir-impac,
+  Title = {A Multiregional Impact Assessment Model for Disaster Analysis},
+  Author = {Elco E. Koks and Mark Thissen},
+  DOI = {10.1080/09535314.2016.1232701},
+  JournalTitle = {Economic Systems Research},
+  Number = {4},
+  Pages = {429-449},
+  URL = {https://doi.org/10.1080/09535314.2016.1232701},
+  Volume = {28},
+  Year = {2016},
+  date_added = {Tue Feb 16 11:06:17 2021},
+}
+
+@article{colon-2020-critic-analy,
+  Title = {Criticality Analysis of a Country's Transport Network Via an Agent-Based Supply Chain Model},
+  Author = {C{\'e}lian Colon and St{\'e}phane Hallegatte and Julie Rozenberg},
+  DOI = {10.1038/s41893-020-00649-4},
+  Journal = {Nature Sustainability},
+  Number = {3},
+  Pages = {209-215},
+  URL = {http://dx.doi.org/10.1038/s41893-020-00649-4},
+  Volume = {4},
+  Year = {2020},
+  date_added = {Fri Dec 8 11:06:46 2023},
+}
+
+@article{otto-2017-model-loss,
+  Title = {Modeling Loss-Propagation in the Global Supply Network: the Dynamic Agent-Based Model Acclimate},
+  Author = {C. Otto and S.N. Willner and L. Wenz and K. Frieler and A. Levermann},
+  DOI = {10.1016/j.jedc.2017.08.001},
+  JournalTitle = {Journal of Economic Dynamics and Control},
+  Number = {nil},
+  Pages = {232--269},
+  URL = {http://dx.doi.org/10.1016/j.jedc.2017.08.001},
+  Volume = {83},
+  Year = {2017},
+  date_added = {Mon Apr 11 11:25:16 2022},
+}
+
+@article{shughrue-2020-global-spread,
+  Title = {Global Spread of Local Cyclone Damages Through Urban Trade Networks},
+  Author = {Chris Shughrue and BT Werner and Karen C. Seto},
+  DOI = {10.1038/s41893-020-0523-8},
+  JournalTitle = {Nature Sustainability},
+  Number = {8},
+  Pages = {606-613},
+  URL = {http://dx.doi.org/10.1038/s41893-020-0523-8},
+  Volume = {3},
+  Year = {2020},
+  date_added = {Wed Jun 15 12:00:32 2022},
+}
+
+@article{shughrue-2018-system-vulner,
+  Title = {Systemic Vulnerabilities of the Global Urban-Industrial Network To Hazards},
+  Author = {Chris Shughrue and Karen C. Seto},
+  DOI = {10.1007/s10584-018-2293-0},
+  JournalTitle = {Climatic Change},
+  Number = {2},
+  Pages = {173-187},
+  URL = {http://dx.doi.org/10.1007/s10584-018-2293-0},
+  Volume = {151},
+  Year = {2018},
+  date_added = {Wed Jun 15 11:59:20 2022},
+}
