+++ boario/event.py
@@ -62,16 +62,31 @@ LOW_DEMAND_THRESH = 10
 
 
 class Event(ABC):
-    # Class Attributes
+    r"""An Event object stores all information about a unique shock during simulation
+    such as time of occurrence, duration, type of shock, amount of damages.
+    Computation of recovery or initially requested rebuilding demand is also
+    done in this class.
+
+    .. warning::
+       The Event class is abstract and cannot be instantiated directly. Only its non-abstract subclasses can be instantiated.
+
+    .. note::
+       Events should be constructed using :py:meth:`~Event.from_series()`, :py:meth:`~Event.from_dataframe()`, :py:meth:`~Event.from_scalar_industries()` or from :py:meth:`~Event.from_scalar_regions_sectors()`.
+       Depending on the type of event chosen, these constructors require additional keyword arguments, that are documented for each instantiable Event subclass.
+       For instance, :py:class:`EventKapitalRebuild` additionally requires `rebuild_tau` and `rebuilding_sectors`.
+
+    .. seealso::
+       Tutorial :ref:`boario-events`
+    """
 
     possible_sectors: pd.Index = pd.Index([])
-    r"""List of sectors present in the MRIO used by the model"""
+    r"""List of sectors present in the MRIOT used by the model"""
 
     possible_regions: pd.Index = pd.Index([])
-    r"""List of regions present in the MRIO used by the model"""
+    r"""List of regions present in the MRIOT used by the model"""
 
     possible_final_demand_cat: pd.Index = pd.Index([])
-    r"""List of final demand categories present in the MRIO used by the model"""
+    r"""List of final demand categories present in the MRIOT used by the model"""
 
     temporal_unit_range: int = 0
     r"""Maximum temporal unit simulated"""
@@ -92,7 +107,7 @@ class Event(ABC):
     r"""lexicographic sector indexes"""
 
     model_monetary_factor: int = 1
-    r"""Amount of unitary currency used in the MRIO (e.g. 1000000 if in € millions)"""
+    r"""Amount of unitary currency used in the MRIOT (e.g. 1000000 if in € millions)"""
 
     gva_df: pd.Series = pd.Series([], dtype="float64")
     r"""GVA per (region,sector)"""
@@ -107,7 +122,7 @@ class Event(ABC):
     r"""Normalized final consumption matrix"""
 
     mrio_name: str = ""
-    r"""MRIO identification"""
+    r"""MRIOT identification"""
 
     @abstractmethod
     def __init__(
@@ -118,24 +133,6 @@ class Event(ABC):
         occurrence: int = 1,
         duration: int = 1,
     ) -> None:
-        r"""Create an event shocking the model from a dictionary.
-
-        An Event object stores all information about a unique shock during simulation such as
-        time of occurrence, duration, type of shock, amount of damages. Computation
-        of recovery or initially requested rebuilding demand is also done in this
-        class.
-
-        Parameters
-        ----------
-
-        event : dict
-            A dictionary holding the necessary information to define an event.
-
-        Examples
-        --------
-            FIXME: Add docs.
-
-        """
         logger.info("Initializing new Event")
         logger.debug("Checking required Class attributes are defined")
 
@@ -206,19 +203,49 @@ class Event(ABC):
         name: Optional[str] = None,
         **kwarg,
     ) -> Event:
-        """Create an event from an impact pd.Series.
+        """Create an event for an impact given as a pd.Series.
 
-        Args:
-            impact (pd.Series): A vector definition of the impact per (region,sector)
-            occurrence (int, optional): The ordinal of occurrence of the event (requires to be > 0). Defaults to 1.
-            duration (int, optional): The duration of the event (entire impact applied during this number of steps). Defaults to 1.
[…]
