+++ docs/source/_autosummary/boario.extended_models.rst
@@ -1,0 +1,31 @@
+extended\_models
+================
+
+.. automodule:: boario.extended_models
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+      :toctree:
+      :template: custom-class-template.rst
+   
+      ARIOPsiModel
+   
+   
+
+   
+   
+   
+
+
+

[…]
