+++ .github/workflows/static_html_deploy.yml
@@ -4,7 +4,7 @@ name: Build and deploy static content to Pages
 on:
   # Runs on pushes targeting the default branch
   push:
-    branches: ["master"]
+    branches: ["main"]
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
