+++ pyproject.toml
@@ -11,7 +11,7 @@ python=">=3.9,<4.0"
 coloredlogs = "^15.0.1"
 country-converter = "^1.0.0"
 dask = ">=2023"
-numpy = "<1.24"
+numpy = "<1.27"
 odfpy = "^1.4.1"
 pandas = ">=1.5"
 progressbar2 = "^4.2.0"
