### 2024-04-22 32827a5
event: kwargs in instantiation
Quick fix for e.g. calling EventKapitalRebuild.from_X(corrects arguments,
recovery_time), which notably happens in the coupling with climada.
A better fix could emerge from #110.
