### 2024-04-22 be2c4ce
event: rebuild_tau as property
Protects against impossible values of type.
