### 2024-04-22 02f462e
event: Additional tests (EventKapitalRebuild)
