### 2024-04-22 e5c4dcb
event: fixes bug when all regions are affected
In this case self._rebuilding_industries_RoW_idx is empty
which caused problems down the road.
