+++ .github/workflows/static_html_deploy.yml
@@ -24,6 +24,7 @@ jobs:
   # Single deploy job since we're just deploying
    build:
     runs-on: ubuntu-latest
+    environment: github-pages
     steps:
       - uses: actions/checkout@v4
         with:
