+++ .github/pull_request_template.md
@@ -1,0 +1,15 @@
+Changes proposed in this PR:
+- 
+- 
+
+This PR fixes #
+
+### PR Author Checklist
+
+- [ ] If new release, is version bumped?
+- [ ] Correct target branch selected (if unsure, select `develop`)
+- [ ] Descriptive pull request title added
+- [ ] Source branch up-to-date with target branch
+- [ ] Documentation updated
+- [ ] Tests passing
+- [ ] [Changelog](https://github.com/spjuhel/blob/main/CHANGELOG.md) updated
