+++ .github/pull_request_template.md
@@ -12,4 +12,4 @@ This PR fixes #
 - [ ] Source branch up-to-date with target branch
 - [ ] Documentation updated
 - [ ] Tests passing
-- [ ] [Changelog](https://github.com/spjuhel/blob/main/CHANGELOG.md) updated
+- [ ] [Changelog](https://github.com/spjuhel/blob/main/CHANGELOG.rst) updated
