+++ pyproject.toml
@@ -20,7 +20,7 @@ pymrio = ">=0.4"
 
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^7.3.1"
+pytest = ">=7.3.1,<9.0.0"
 pygit2 = "^1.12.1"
 ipykernel = "^6.23.0"
 black = ">=23.3,<25.0"
