+++ pyproject.toml
@@ -30,7 +30,7 @@ sphinx-automodapi = "^0.15.0"
 sphinx-copybutton = "^0.5.2"
 pydata-sphinx-theme = ">=0.13.3,<0.16.0"
 nbsphinx = "^0.9.2"
-pytest-cov = "^4.1.0"
+pytest-cov = ">=4.1,<6.0"
 isort = "^5.13.2"
 flake8 = "^7.0.0"
 flake8-pyproject = "^1.2.3"
