+++ docs/source/versions.yaml
@@ -1,5 +1,5 @@
 "latest":
-  tag: '0.5.8'
+  tag: 'v0.5.8'
   languages:
     - "en"
 
