### 2024-04-23 f39a767
quick fix for build_docs.py
