+++ docs/source/versions.yaml
@@ -1,4 +1,4 @@
-"latest":
+"v0.5.8":
   tag: 'v0.5.8'
   languages:
     - "en"
