### 2024-04-26 #6547 @mwt
@spjuhel, this is a big improvement. I really like the new documentation.

I was running version 0.5.7 and am now on 0.5.9. I no longer get a key error, but the function still fails to converge when I use numerical shares instead of gdp. Perhaps this was fixed though in a version that isn't yet on pypi? Here is my example. For me, there are NaN for every step after 1.

```py
# import pymrio for the test MRIO
import pymrio

# import the different classes
from boario.simulation import Simulation  # Simulation wraps the model
from boario.extended_models import ARIOPsiModel  # The core of the model

from boario.event import EventKapitalRebuild  # A class defining a shock on capital

# Load the IOSystem from pymrio
mrio = pymrio.load_test().calc_all()

# Instantiate the model and the simulation
model = ARIOPsiModel(mrio)
sim = Simulation(model,n_temporal_units_to_sim=730)

# Instantiate an event.
ev = EventKapitalRebuild.from_scalar_regions_sectors(
  impact=500000,
  regions=["reg1"],
  sectors=["manufactoring", "mining"],
  impact_sectoral_distrib = [0.5, 0.5],
  rebuilding_sectors={"construction": 0.55,"manufactoring": 0.45},
  rebuilding_factor=1.0,
  rebuild_tau=90,
)

# Add the event to the simulation
sim.add_event(ev)

# Launch the simulation
sim.loop(progress=False)

# You should be able to generate a dataframe of
# the production with the following line
df = sim.production_realised
# This allows to normalize production at its initial level
df = df / df.loc[0]

df.loc[:, ("reg1", slice(None))].plot()
```
I think you addressed everything else. The "See Also" links in the function documentation that lead to tutorials is a nice touch.
