### 2024-04-29 9575319
Bump sphinx-automodapi from 0.15.0 to 0.17.0
Bumps [sphinx-automodapi](https://github.com/astropy/sphinx-automodapi) from 0.15.0 to 0.17.0.
- [Release notes](https://github.com/astropy/sphinx-automodapi/releases)
- [Changelog](https://github.com/astropy/sphinx-automodapi/blob/main/CHANGES.rst)
- [Commits](https://github.com/astropy/sphinx-automodapi/compare/v0.15.0...v0.17.0)

---
updated-dependencies:
- dependency-name: sphinx-automodapi
  dependency-type: direct:development
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
