### 2024-04-29 #6547 @spjuhel
@mwt Thanks for noticing these issues,

For the example, I realize that the shock is just too great on the mining sector with this setup, and that make the model "crash", but thus it should at least raise a proper warning. And a functioning example would be more appropriate.

> I think you should raise an error in this case? That would cause the test to fail.
> Do you know why the impact has size zero in this test case?

I realize the test/code is not very well-designed, I don't have a clear method for instantiating from a numpy array, and in the test this case (empty_np, where input is an empty numpy array) goes to the `from_scalar_regions_sectors()` which assumes the impact is scalar, hence the warning. The test does raise an error afterwards, as it should, but that could be made more proper and clearer indeed.

> I'm trying to understand if there are other inputs and edge cases where it fails.

That would be so helpful, I did look out for these, but testing is not my forte and there's a lot of room for improvements here.

------
I actually plan (long term) to highly simplify the event module, such that:

- event instantiation will be done via `event.from_x(impact, event_type, ...)` (module functions instead of class methods), which will call the correct class instantiation method.
- all the class attributes that check the validity will be removed from the module and checks will be done in the simulation context instead. At the moment, you cannot instantiate events without first instantiating a model and a simulation, which sets those class attributes. Looking back, this was a terrible idea...

But these require to redesign technical parts of all modules, so I'm not sure if they should be considered in this review?
(Any remarks/suggestions on this are most welcome!!)
