### 2024-05-15 #6547 @potterzot
I was able to install and reproduce the example models as well as successfully run tests. No issues there. I do have several suggestions on the documentation and for the paper. I've also made a [PR](https://github.com/spjuhel/BoARIO/pull/119) with some small grammar fixes. I've checked off all of the items in the review checklist, but I think the package would benefit from addressing some of the comments below before publication. Most of these are very minor.

This is beyond the scope of the current paper, but given the goal of the project stated in both the README and the paper, it would be useful to have more in depth case studies that demonstrate both how to conduct various real-world analysis of impacts and how to modify the base assumptions and data used in the model. I think this would substantially help with adoption of this software for use in impact modeling by other researchers and impact modelers. A few well-designed examples would both highlight how users should think about measuring impacts of different kinds of shocks as well as how to modify the model parameters to reflect changes. In addition, an example of how to take an IO table and create a pymrio object from it would allow a user who doesn't have access to a prebuild IO model but does have access to their country's make and use tables to build their own underlying model. I know I've wanted something like this to exist for quite some time. In the US there are some efforts to provide an open source IO model to "democratize" the use of impact modeling in policy advocacy (for example, see [Tapestry](https://www.uidaho.edu/cals/tapestry)).

There are some integrations and events that are referenced in the documentation and paper but that aren't currently working. For example, one of the three possible impact events ([EventArbitraryProd](https://spjuhel.github.io/BoARIO/_autosummary/boario.event.EventArbitraryProd.html#boario.event.EventArbitraryProd)) is currently not available because of a critical bug. The paper implies a fully working model that has extensive documentation and case studies, but there aren't case studies or multiple types of shocks described or linked in the documentation or the paper. The paper also states that integration to various other modeling platforms and workflows (e.g. "shock on demand, shock on production, shock on both, shock involving reconstruction or not, etc" suggests shock possibilities that are not described in the documentation. Mostly this could be addressed by toning down the promise of `boario` a bit. I make a few specific suggestions in the comments below.

I hope the following comments are helpful. Some of the comments are very minor. Probably they do not all need to be addressed, but I think they all would improve the overall sense of "completeness" of the software and the documentation.

## Documentation comments
- [README](https://github.com/spjuhel/BoARIO/blob/main/README.rst): The `Hallegatte 2008` citation is not linked as the other citations are.
- [README](https://github.com/spjuhel/BoARIO/blob/main/README.rst): It might be nice for the new user who is not familiar with these specific IO models to link to EXIOBASE3 and EORA26.
- [README](https://github.com/spjuhel/BoARIO/blob/main/README.rst): being familiar with several IO models in the United States, but not with pymrio or the models mentioned in the README, would it be possible to add some more information about what kind of object a pymrio object is? I've made a small edit in the linked PR. There is also great documentation ([example](https://spjuhel.github.io/BoARIO/boario-math.html)) that describes the model, but it is not linked in the README. I think adding a link to this kind of documentation where it might help a new user who is familiar with IO models generally but not with this project specifically understand some of the capabilities of this model.
- [Quickstart](https://spjuhel.github.io/BoARIO/notebooks/boario-quickstart.html): ""Some attributes of the test MRIOT are not computed. Calling calc_all() ensures all required tables are present in the IOSystem object." It would be nice here to know a little bit more specifically what `calc_all()` is doing. If the MRIOT is incomplete, is `calc_all()` using multiple imputation to fill missing values in the IO table? I understand that I can look at the function API reference, but describing what it is doing in the quickstart would also be helpful.
- [Model Parameters](https://spjuhel.github.io/BoARIO/tutorials/model-params.html): There are some differences in how citations are formatted in the different documents. In the README, it's "(Halegatte 2013)". In the tutorial documentation, it's "[Hal13]". Then in the [model description](https://spjuhel.github.io/BoARIO/boario-math.html) it is back to "Hallegatte 2014" and does not link to a reference. I would make these all follow the same citation style and all link to the actual reference if possible.

## Paper comments
- [Summary](https://github.com/spjuhel/BoARIO/blob/main/paper.md#summary): "The impacts of economic shocks (caused by natural or technological disasters for instance) often extend far beyond the cost of their local, direct consequences, as the economic perturbations they cause propagate along supply chains." This seems limiting, since perturbations propagate along both supply and demand vectors, and have consequences in environmental and social well being dimensions.
- [Statement of Need](https://github.com/spjuhel/BoARIO/blob/main/paper.md#statement-of-need): "there were used for" should be "they were used for"
- [Statement of Need](https://github.com/spjuhel/BoARIO/blob/main/paper.md#statement-of-need): "accompanied by the extensive [online documentation](https://spjuhel.github.io/BoARIO/) (where a more in depth description is available)" I think there is the potential for this to be true, but currently I think this is a little overstated. I would remove "extensive" for the time being until case studies and explanations of why and how different impact estimates may be conducted have been added.
- [Statement of Need](https://github.com/spjuhel/BoARIO/blob/main/paper.md#statement-of-need): "Other notable ongoing projects, are" I would either expand on the last two of these or remove them. Are there working papers or code repositories available for these projects? As stated, they are ambiguous.
- [Status](https://github.com/spjuhel/BoARIO/blob/main/paper.md#status): "Although its current version is fully operational, further improvements, notably the implementation of additional economic mechanisms or variations of existing ones are already planned." Given that one of the event impacts is not operational I would restate this or remove the part preceding "further improvements ..."
- [Status](https://github.com/spjuhel/BoARIO/blob/main/paper.md#status): "Integration tests can be run using pytest" should have a period at the end.
