+++ boario/event.py
@@ -16,17 +16,14 @@
 from __future__ import annotations
 
 import inspect
-import math
 import warnings
 from abc import ABC, abstractmethod
-from functools import partial
-from typing import Callable, List, Optional, Tuple, Union
+from typing import Callable, List, Literal, Optional, Tuple, Union
 
 import numpy as np
 import numpy.typing as npt
 import pandas as pd
 from pandas.api.types import is_numeric_dtype
-
 
 from boario import logger
 from boario.utils.recovery_functions import (
@@ -62,6 +59,277 @@ REBUILDING_FINALDEMAND_CAT_REGEX = (
 
 LOW_DEMAND_THRESH = 10
 
+def from_series(
+    impact: pd.Series,
+    *,
+    event_type: str,
+    occurrence: int = 1,
+    duration: int = 1,
+    name: Optional[str] = None,
+    event_monetary_factor: int | None = None,
+    recovery_tau: int | None = None,
+    recovery_function: str | None = "linear",
+    households_impact: Impact | None = None,
+    rebuild_tau: int | None = None,
+    rebuilding_sectors: dict[str, float] | pd.Series | None = None,
+    rebuilding_factor: float | None = 1.0,
+) -> Event:
+    """Create an event for an impact given as a pd.Series.
+
+    Parameters
+    ----------
+    impact : pd.Series
+        A pd.Series defining the impact per (region, sector)
+    occurrence : int
+        The ordinal of occurrence of the event (requires to be > 0). Defaults to 1.
+    duration : int
+        The duration of the event (entire impact applied during this number of steps). Defaults to 1.
+    name : Optional[str]
+        A possible name for the event, for convenience. Defaults to None.
+    **kwargs :
+        Keyword arguments keyword arguments to pass to the instantiating method
+    (depends on the type of event).
+
+    Returns
+    -------
+    Event
+        An Event object or one of its subclass
+
+    Raises
+    ------
+    ValueError
+        Raised if impact is empty of contains negative values.
+
+    Examples
+    --------
+    >>> import pandas as pd
+    >>> import pymrio as pym
+    >>> from boario.simulation import Simulation
+    >>> from boario.extended_model import ARIOPsiModel
+    >>> from boario.event import EventKapitalRecover
+    >>>
+    >>> mriot = pym.load_test()
+    >>> mriot.calc_all()
+    >>>
+    >>> impact_series = pd.Series({('reg1', 'electricity'): 100000.0, ('reg1', 'mining'): 50000.0})
+    >>> model = ARIOPsiModel(mriot)
+    >>> sim = Simulation(model)
+    >>> event = EventKapitalRecover.from_series(impact_series, occurrence=5, duration=10, recovery_time=30, name="Event 1")
+    >>> sim.add_event(event)
+
+    """
+
+    if event_type == "rebuild":
+        return EventKapitalRebuild._from_series(
+            impact=impact,
+            occurrence=occurrence,
+            duration=duration,
+            name=name,
+            event_monetary_factor=event_monetary_factor,
+            households_impact=households_impact,
+            rebuild_tau=rebuild_tau,
+            rebuilding_sectors=rebuilding_sectors,
+            rebuilding_factor=rebuilding_factor,
+        )
+    elif event_type == "recovery":
+        return EventKapitalRecover._from_series(
+            impact=impact,
+            occurrence=occurrence,
+            duration=duration,
+            name=name,
+            event_monetary_factor=event_monetary_factor,
+            recovery_tau=recovery_tau,
+            recovery_function=recovery_function,
+        )
+    elif event_type == "arbitrary":
+        raise NotImplementedError("This type of event is not implemented yet.")
+
+    else:
[…]
