+++ boario/model_base.py
@@ -28,13 +28,7 @@ import numpy.typing as npt
 import pandas as pd
 from pymrio.core.mriosystem import IOSystem
 
-from boario import logger, warn_once
-from boario.event import (
-    Event,
-    EventArbitraryProd,
-    EventKapitalDestroyed,
-    EventKapitalRebuild,
-)
+from boario import logger
 from boario.utils.misc import lexico_reindex
 
 __all__ = [
@@ -80,10 +74,111 @@ VA_idx = np.array(
 )
 
 
+def _fast_sum(array: np.ndarray, axis: int) -> np.ndarray:
+    """
+    Perform a fast summation over the specified axis of the input array using einsum.
+
+    Parameters:
+        array (numpy.ndarray): Input array.
+        axis (int): Axis along which to perform the summation.
+
+    Returns:
+        numpy.ndarray: Summed array along the specified axis.
+
+    Raises:
+        ValueError: If the specified axis is out of bounds for the input array.
+
+    Example:
+        array = np.array([[1, 2, 3], [4, 5, 6]])
+        result = _fast_sum(array, 0)  # Sum along the first axis
+        print(result)  # Output: [5 7 9]
+    """
+    if axis == 0:
+        if array.ndim == 1:
+            return np.einsum("i->", array)
+        if array.ndim == 2:
+            return np.einsum("ij->j", array)
+        if array.ndim == 3:
+            return np.einsum("ijk->jk", array)
+    elif axis == 1:
+        if array.ndim == 2:
+            return np.einsum("ij->i", array)
+        if array.ndim == 3:
+            return np.einsum("ijk->ik", array)
+    elif axis == 2:
+        return np.einsum("ijk->ij", array)
+    else:
+        raise ValueError("Axis out of bounds for input array.")
+
+
+def _divide_arrays_ignore(a, b):
+    with np.errstate(divide="ignore", invalid="ignore"):
+        ret = np.divide(a, b)
+        np.nan_to_num(ret)
+        return ret
+
+
 class ARIOBaseModel:
+    r"""The core of an ARIO model.  Handles the different arrays containing the mrio tables.
+
+    `ARIOBaseModel` wrap all the data and functions used in the core of the most basic version of the ARIO
+    model (based on :cite:`2013:hallegatte` and :cite:`2020:guan`).
+
+    An ARIOBaseModel instance based is build on the given IOSystem.
+    Most default parameters are the same as in :cite:`2013:hallegatte` and
+    default order mechanisms comes from :cite:`2020:guan`. By default each
+    industry capital stock is 4 times its value added (:cite:`2008:hallegatte`).
+
+    Parameters
+    ----------
+    pym_mrio : IOSystem
+        The IOSystem to base the model on.
+    order_type : str, default "alt"
+        The type of orders mechanism to use. Currently, can be "alt" or
+        "noalt". See :ref:`boario-math`
+    alpha_base : float, default 1.0
+        Base value of overproduction factor :math:`\alpha^{b}` (Default to 1.0).
+    alpha_max : float, default 1.25
+        Maximum factor of overproduction :math:`\alpha^{\textrm{max}}` (default should be 1.25).
+    alpha_tau : int, default 365
+        Characteristic time of overproduction :math:`\tau_{\alpha}` in ``n_temporal_units_by_step`` (default should be 365 days).
+    rebuild_tau : int, default 60
+        Rebuilding characteristic time :math:`\tau_{\textrm{REBUILD}}` (see :ref:`boario-math`). Overwritten by per event value if it exists.
+    main_inv_dur : int, default 90
+        The default numbers of days for inputs inventory to use if it is not defined for an input.
+    monetary_factor : int, default 10**6
[…]

[…]
