### 2024-05-17 32448fb
tests for event module
   - Also added option for pytest coverage
   - Also added pandas-stubs to dev dependencies
   - Also minor changes to event modules (type hints, some error raising and prevention)
