+++ boario/event.py
@@ -18,7 +18,7 @@ from __future__ import annotations
 import inspect
 import warnings
 from abc import ABC, abstractmethod
-from typing import Callable, List, Literal, Optional, Tuple, Union
+from typing import Callable, List, Literal, Optional, Tuple, Union, overload
 
 import numpy as np
 import numpy.typing as npt
@@ -46,7 +46,7 @@ __all__ = [
 ]
 
 VectorImpact = Union[list, dict, np.ndarray, pd.DataFrame, pd.Series]
-ScalarImpact = Union[int, float, np.integer]
+ScalarImpact = int | float
 Impact = Union[VectorImpact, ScalarImpact]
 IndustriesList = Union[List[Tuple[str, str]], pd.MultiIndex]
 SectorsList = Union[List[str], pd.Index, str]
@@ -59,10 +59,44 @@ REBUILDING_FINALDEMAND_CAT_REGEX = (
 
 LOW_DEMAND_THRESH = 10
 
+@overload
 def from_series(
     impact: pd.Series,
     *,
-    event_type: str,
+    event_type: Literal["recovery"],
+    occurrence: int = 1,
+    duration: int = 1,
+    name: Optional[str] = None,
+    event_monetary_factor: int | None = None,
+    recovery_tau: int | None = None,
+    recovery_function: str | None = "linear",
+    households_impact: Impact | None = None,
+    rebuild_tau: int | None = None,
+    rebuilding_sectors: dict[str, float] | pd.Series | None = None,
+    rebuilding_factor: float | None = 1.0,
+) -> EventKapitalRecover: ...
+
+@overload
+def from_series(
+    impact: pd.Series,
+    *,
+    event_type: Literal["rebuild"],
+    occurrence: int = 1,
+    duration: int = 1,
+    name: Optional[str] = None,
+    event_monetary_factor: int | None = None,
+    recovery_tau: int | None = None,
+    recovery_function: str | None = "linear",
+    households_impact: Impact | None = None,
+    rebuild_tau: int | None = None,
+    rebuilding_sectors: dict[str, float] | pd.Series | None = None,
+    rebuilding_factor: float | None = 1.0,
+) -> EventKapitalRebuild: ...
+
+def from_series(
+    impact: pd.Series,
+    *,
+    event_type: Literal["recovery"] | Literal["rebuild"],
     occurrence: int = 1,
     duration: int = 1,
     name: Optional[str] = None,
@@ -138,6 +172,7 @@ def from_series(
             duration=duration,
             name=name,
             event_monetary_factor=event_monetary_factor,
+            households_impact=households_impact,
             recovery_tau=recovery_tau,
             recovery_function=recovery_function,
         )
@@ -158,11 +193,11 @@ def from_scalar_industries(
     duration: int = 1,
     name: Optional[str] = None,
     event_monetary_factor: int | None = None,
-    recovery_tau: int | None = 1,
+    recovery_tau: int | None = None,
     recovery_function: str | None = "linear",
     households_impact: Impact | None = None,
     rebuild_tau: int | None = None,
-    rebuilding_sectors: dict[str, float] | pd.Series,
+    rebuilding_sectors: dict[str, float] | pd.Series | None = None,
     rebuilding_factor: float | None = 1.0,
 ) -> Event:
     """Creates an Event from a scalar and a list of industries affected.
@@ -248,11 +283,11 @@ def from_scalar_regions_sectors(
     duration: int = 1,
     name: Optional[str] = None,
     event_monetary_factor: int | None = None,
-    recovery_tau: int | None = 1,
+    recovery_tau: int | None = None,
     recovery_function: str | None = "linear",
     households_impact: Impact | None = None,
[…]
