+++ tests/test_events.py
@@ -1,7 +1,7 @@
+from boario.utils.recovery_functions import concave_recovery, convexe_recovery, convexe_recovery_scaled, linear_recovery
 import pytest
 
-# import pymrio for the test MRIO
-import pymrio
+from contextlib import nullcontext as does_not_raise
 
 # import pandas for the plot
 import pandas as pd
@@ -10,547 +10,580 @@ import numpy as np
 from numpy.core._exceptions import UFuncTypeError
 
 # import the different classes
-import boario
-from boario.model_base import ARIOBaseModel
-from boario.simulation import Simulation  # Simulation wraps the model
-from boario.event import (
-    Event,
-    EventKapitalRebuild,
-    EventArbitraryProd,
-    EventKapitalRecover,
-)  # A class defining a shock on capital
+import boario.event as event
+
+import pandas as pd
+import pytest
 
 
 @pytest.fixture
-def test_mrio():
-    mrio = pymrio.load_test()  # .calc_all()
-    mrio.aggregate(
-        region_agg=["reg1", "reg1", "reg2", "reg2", "reg3", "reg3"],
-        sector_agg=[
-            "food",
-            "mining",
-            "manufactoring",
-            "other",
-            "construction",
-            "other",
-            "other",
-            "other",
+def sample_series():
+    return pd.Series(
+        {
+            ("RegionA", "Sector1"): 30000.0,
+            ("RegionA", "Sector2"): 10000.0,
+            ("RegionB", "Sector3"): 40000.0,
+        }
+    )
+
+
+class TestEventInitSeries:
+
+    @pytest.mark.parametrize("ev_t", ["recovery", "rebuild"])
+    def test_event_from_series_invalid(self, ev_t):
+        with pytest.raises(ValueError):
+            event.from_series(pd.Series(), event_type=ev_t)
+
+        with pytest.raises(ValueError):
+            event.from_series(pd.Series([0]), event_type=ev_t)
+
+        with pytest.raises(ValueError):
+            event.from_series(pd.Series([-1]), event_type=ev_t)
+
+    def test_event_from_series_recovery(self, sample_series):
+        event_instance = event.from_series(
+            sample_series, event_type="recovery", recovery_tau=1
+        )
+        assert event_instance.occurrence == 1
+        assert event_instance.duration == 1
+        assert event_instance.total_impact == 80000.0
+        assert event_instance.name is None
+        assert event_instance.recovery_function == event.linear_recovery
+        pd.testing.assert_index_equal(
+            event_instance.aff_industries, sample_series.index, check_names=False
+        )
+        pd.testing.assert_index_equal(
+            event_instance.aff_regions,
+            pd.Index(["RegionA", "RegionB"]),
+            check_names=False,
+        )
+        pd.testing.assert_index_equal(
+            event_instance.aff_sectors,
+            pd.Index(["Sector1", "Sector2", "Sector3"]),
+            check_names=False,
+        )
+        pd.testing.assert_series_equal(
+            event_instance.impact, sample_series, check_names=False
+        )
+        pd.testing.assert_series_equal(
+            event_instance.impact_industries_distrib,
+            pd.Series(
+                {
+                    ("RegionA", "Sector1"): 3 / 8,
+                    ("RegionA", "Sector2"): 1 / 8,
+                    ("RegionB", "Sector3"): 4 / 8,
+                }
+            ),
+            check_names=False,
+        )
+        pd.testing.assert_series_equal(
+            event_instance.impact_regional_distrib,
+            pd.Series({"RegionA": 1 / 2, "RegionB": 1 / 2}),
+            check_names=False,
+        )
+
[…]
