+++ boario/simulation.py
@@ -596,7 +596,8 @@ class Simulation:
             self.equi[(n_checks, self.current_temporal_unit, "stocks")] = "not equi"
 
         if (
-            self.model.rebuild_demand_tot is None or self.model.rebuild_demand is None
+            self.model.rebuild_demand_tot is None
+            or self.model.rebuild_demand is None
             or not self.model.rebuild_demand.any()
         ):
             self.equi[(n_checks, self.current_temporal_unit, "rebuilding")] = "finished"
@@ -684,7 +685,7 @@ class Simulation:
         rebuild/recover)
 
         """
-        new_reb_event=False
+        new_reb_event = False
         for event_tracker in self._event_tracking:
             if event_tracker.status == "pending":
                 if (
@@ -701,7 +702,9 @@ class Simulation:
                     event_tracker._status = "happening"
         for event_tracker in self._event_tracking:
             if event_tracker.status == "happening":
-                if self.current_temporal_unit >= (event_tracker.event.occurrence + event_tracker.event.duration):
+                if self.current_temporal_unit >= (
+                    event_tracker.event.occurrence + event_tracker.event.duration
+                ):
                     if isinstance(event_tracker.event, EventKapitalRebuild):
                         new_reb_event = True
                         self.model._n_rebuilding_events += 1
@@ -717,7 +720,9 @@ class Simulation:
                             )
                         )
 
-                    elif isinstance(event_tracker, (EventKapitalRecover, EventArbitraryProd)):
+                    elif isinstance(
+                        event_tracker, (EventKapitalRecover, EventArbitraryProd)
+                    ):
                         event_tracker._status = "recovering"
                         logger.info(
                             "Temporal_Unit : {} ~ Event named {} that occurred at {} in {} for {} damages has started recovering".format(
@@ -750,34 +755,42 @@ class Simulation:
         for event_tracker in self._event_tracking:
             if event_tracker.status == "rebuilding":
                 if event_tracker._rebuild_id is None:
-                    raise RuntimeError("Rebuilding event has no rebuilding id, which should not happen.")
+                    raise RuntimeError(
+                        "Rebuilding event has no rebuilding id, which should not happen."
+                    )
                 assert self.model.rebuild_prod is not None
-                event_tracker.receive_indus_rebuilding(self.model.rebuild_prod_indus_event(event_tracker._rebuild_id))
+                event_tracker.receive_indus_rebuilding(
+                    self.model.rebuild_prod_indus_event(event_tracker._rebuild_id)
+                )
 
                 if event_tracker.rebuild_demand_house is not None:
-                    event_tracker.receive_indus_rebuilding(self.model.rebuild_prod_house_event(event_tracker._rebuild_id))
-                if event_tracker._house_dmg is None and event_tracker._indus_dmg is None:
-                    event_tracker._status="finished"
+                    event_tracker.receive_indus_rebuilding(
+                        self.model.rebuild_prod_house_event(event_tracker._rebuild_id)
+                    )
+                if (
+                    event_tracker._house_dmg is None
+                    and event_tracker._indus_dmg is None
+                ):
+                    event_tracker._status = "finished"
                     events_rebuilt_ids.append(event_tracker._rebuild_id)
                     self._events_to_rebuild -= 1
 
         if len(events_rebuilt_ids) > 0:
             events_rebuilt_ids = sorted(events_rebuilt_ids)
[…]
