+++ boario/event.py
@@ -59,6 +59,7 @@ REBUILDING_FINALDEMAND_CAT_REGEX = (
 
 LOW_DEMAND_THRESH = 10
 
+
 @overload
 def from_series(
     impact: pd.Series,
@@ -76,6 +77,7 @@ def from_series(
     rebuilding_factor: float | None = 1.0,
 ) -> EventKapitalRecover: ...
 
+
 @overload
 def from_series(
     impact: pd.Series,
@@ -92,6 +94,7 @@ def from_series(
     rebuilding_sectors: dict[str, float] | pd.Series | None = None,
     rebuilding_factor: float | None = 1.0,
 ) -> EventKapitalRebuild: ...
+
 
 def from_series(
     impact: pd.Series,
@@ -490,7 +493,9 @@ class Event(ABC):
     @classmethod
     def _distribute_impact(cls, impact_vec: pd.Series, distrib: pd.Series) -> pd.Series:
         if not isinstance(impact_vec, pd.Series):
-            raise ValueError(f"Impact vector has to be a Series not a {type(impact_vec)}.")
+            raise ValueError(
+                f"Impact vector has to be a Series not a {type(impact_vec)}."
+            )
         if impact_vec.size < 1:
             raise ValueError(f"Impact vector cannot be null sized.")
         if distrib.sum() != 1.0:
@@ -498,7 +503,9 @@ class Event(ABC):
 
         ret = impact_vec * distrib
         if ret.hasnans:
-            raise ValueError("Products of impact vector and distrib lead to NaNs, check index matching and values.")
+            raise ValueError(
+                "Products of impact vector and distrib lead to NaNs, check index matching and values."
+            )
         return ret
 
     @classmethod
@@ -518,8 +525,12 @@ class Event(ABC):
         affected_industries = cls._build_industries_idx(
             affected_regions, affected_sectors
         )
-        regional_distrib = cls._level_distrib(affected_industries.levels[0], impact_regional_distrib)
-        sectoral_distrib = cls._level_distrib(affected_industries.levels[1], impact_sectoral_distrib)
+        regional_distrib = cls._level_distrib(
+            affected_industries.levels[0], impact_regional_distrib
+        )
+        sectoral_distrib = cls._level_distrib(
+            affected_industries.levels[1], impact_sectoral_distrib
+        )
         industries_distrib = pd.Series(
             np.outer(regional_distrib.values, sectoral_distrib.values).flatten(),  # type: ignore
             index=pd.MultiIndex.from_product(
@@ -581,11 +592,13 @@ class Event(ABC):
         affected_idx: List[str] | pd.Index | pd.MultiIndex,
         distrib: Literal["equal"] | pd.Series,
     ):
-        if isinstance(distrib,str) and distrib == "equal":
+        if isinstance(distrib, str) and distrib == "equal":
             return cls._distrib_equi_level(affected_idx)
         else:
             if not isinstance(distrib, pd.Series):
-                raise ValueError("The given impact distribution is incorrect. (Pandas Series required).")
+                raise ValueError(
+                    "The given impact distribution is incorrect. (Pandas Series required)."
+                )
             affected_idx = pd.Index(affected_idx)
             if not affected_idx.isin(distrib.index).all():
                 raise ValueError(
@@ -626,16 +639,16 @@ class Event(ABC):
     @impact.setter
     def impact(self, value: pd.Series):
         self._impact_df = value
-        self._impact_df.rename_axis(index=["region","sector"], inplace=True)
+        self._impact_df.rename_axis(index=["region", "sector"], inplace=True)
         logger.debug("Sorting impact Series")
         self._impact_df.sort_index(inplace=True)
         tmp_idx = self.impact.loc[self.impact > 0].index
         if not isinstance(tmp_idx, pd.MultiIndex):
[…]
