### 2024-05-24 b14bffb
moved _fast_sum and _divide_arrays to utils
