### 2024-05-24 e5f3099
Integrates remarks on paper from N. Potter.
