### 2024-05-24 ce52f6e
Make citation style more consistent
