### 2024-05-24 46661d1
Integrates remarks on README from N. Potter
Also adds a short section on MRIOTs.
