+++ README.rst
@@ -70,8 +70,30 @@ Each economic sector produces its generic product and draws inputs from an inven
 Each sector answers to a total demand consisting of a final demand (household consumption,
 public spending and private investments) of all regions (local demand and exports) and
 intermediate demand (through inputs inventory resupply). An initial equilibrium state of
-the economy is built based on multi-regional input-output tables (MRIO tables).
+the economy is built based on multi-regional input-output tables (MRIOTs).
 
+For a more detailed description, please refer to the `Mathematical documentation`_ of the model.
+
+Multi-Regional Input-Output tables
+-------------------------------------
+
+Multi-Regional Input-Output tables (MRIOTs) are comprehensive economic data sets
+that capture inter-regional trade flows, production activities, and consumption
+patterns across different regions or countries. These tables provide a detailed
+breakdown of the flows of goods and services between industries within each
+region and between regions themselves. MRIOTs are constructed through a
+combination of national or regional input-output tables, international trade
+data, and other relevant economic statistics. By integrating data from multiple
+regions, MRIOTs enable the analysis of global supply chains, international trade
+dependencies, and the estimation of economic impacts across regions. However,
+they also come with limitations, such as data inconsistencies across regions,
+assumptions about trade patterns and production technologies, and the challenge
+of ensuring coherence and accuracy in the aggregation of data from various
+sources.
+
+.. _`Mathematical documentation`: https://spjuhel.github.io/BoARIO/boario-math.html
+
+.. _`Hal08`: https://doi.org/10.1111/risa.12090
 
 Where to get BoARIO ?
 ==========================
@@ -98,7 +120,7 @@ How does BoARIO work?
 
 In a nutshell, BoARIO takes the following inputs :
 
-- an Environmentally Extended IO table (such as EXIOBASE3 or EORA26) in the form of an ``pymrio.IOSystem`` object, using the `Pymrio`_ python package. Please reference the `Pymrio documentation <https://github.com/IndEcol/pymrio>`_ for details on methods available to pymrio objects
+- a (possibly Environmentally Extended) Multi-Regional IO table (such as `EXIOBASE 3`_ or `EORA26`_) in the form of an ``pymrio.IOSystem`` object, using the `Pymrio`_ python package. Please reference the `Pymrio documentation <https://github.com/IndEcol/pymrio>`_ for details on methods available to pymrio objects.
 
 - multiple parameters which govern the simulation,
 
@@ -109,6 +131,9 @@ And produces the following outputs:
 - the step by step, sector by sector, region by region evolution of most of the variables involved in the simulation (`production`, `demand`, `stocks`, ...)
 
 - aggregated indicators for the whole simulation (`shortages duration`, `aggregated impacts`, ...)
+
+.. _`EXIOBASE 3`: https://www.exiobase.eu/
+.. _`EORA26`: https://worldmrio.com/eora26/
 
 Example of use
 =================
