### 2024-05-24 4550920
Tuned down the available events to match current project status
