### 2024-05-24 #6547 @spjuhel
Thanks a lot for all these precious comments,

> it would be useful to have more in depth case studies that demonstrate both how to conduct various real-world analysis of impacts and how to modify the base assumptions and data used in the model. I think this would substantially help with adoption of this software for use in impact modeling by other researchers and impact modelers. A few well-designed examples would both highlight how users should think about measuring impacts of different kinds of shocks as well as how to modify the model parameters to reflect changes.

I totally agree. I intend to extend the documentation with more detailed examples based on my papers, but these are still in the reviewing process, or even unfinished. Another thing I would like to do is to present examples reproducing previous studies / well-known disasters in the indirect impact literature (such as Hurricane Katrina or the 2001 floods in Thailand). This however requires a substantial amount of work, which I'm not at liberty to do at the moment.

> In addition, an example of how to take an IO table and create a pymrio object from it would allow a user who doesn't have access to a prebuild IO model but does have access to their country's make and use tables to build their own underlying model.

I agree this would be a nice addition, but I feel like this is something that should be part of the pymrio documentation ([and already does to some extent](https://pymrio.readthedocs.io/en/latest/notebooks/pymrio_directly_assign_attributes.html)).

> There are some integrations and events that are referenced in the documentation and paper but that aren't currently working. For example, one of the three possible impact events ([EventArbitraryProd](https://spjuhel.github.io/BoARIO/_autosummary/boario.event.EventArbitraryProd.html#boario.event.EventArbitraryProd)) is currently not available because of a critical bug. The paper implies a fully working model that has extensive documentation and case studies, but there aren't case studies or multiple types of shocks described or linked in the documentation or the paper. The paper also states that integration to various other modeling platforms and workflows (e.g. "shock on demand, shock on production, shock on both, shock involving reconstruction or not, etc" suggests shock possibilities that are not described in the documentation. Mostly this could be addressed by toning down the promise of boario a bit. I make a few specific suggestions in the comments below.

Yes, I might have been too ambitious when writing some parts of the paper. I discovered only during this review that `EventArbitraryProd` was not working properly. I am working on a new major version which will include a fix for this, but I also took this opportunity to completely rework the way events are handled (as mentioned before). I have rephrased the corresponding part to make it more consistent with the current implementation.

I just pushed these changed to the `develop` branch, but I still want to look into what @mwt raised and recheck that everything works properly in the examples before merging to main and publishing another release.
