+++ .github/workflows/static_html_deploy.yml
@@ -46,7 +46,9 @@ jobs:
           # which you find below to build all documents
           git fetch --tags
           cd docs
-          poetry run python ./source/build_docs.py
+          poetry shell
+          ipython kernel install --name "boario"
+          python ./source/build_docs.py
       - name: Setup Pages
         uses: actions/configure-pages@v5
       - name: Upload artifact
