### 2024-05-24 c68421d
poetry shell not working on CI
