### 2024-05-30 fc325de
disables console logging in tests
