### 2024-05-30 ff61993
Event: checks impact is scalar in scalar case
