+++ boario/event.py
@@ -494,8 +494,11 @@ class Event(ABC):
         Event
             An Event object or one of its subclass.
         """
+        if not isinstance(impact, (int,float)):
+            raise ValueError("Impact is not scalar.")
+
         if impact <= 0:
-            raise ValueError("Impact is null")
+            raise ValueError("Impact is null.")
 
         if isinstance(regions, str):
             regions = [regions]
