### 2024-05-30 0960286
model: improves message for sectoral impacts being to big
