### 2024-05-30 #6547 @spjuhel
Here are my latest changes on `develop`:

- I've reactivated logging to console by default, so that crashes due to a shock being too important are not silenced by default. (c.f. your case @mwt when setting `impact_sectoral_distrib = [0.5, 0.5]` ). In the future, I will add a check directly when adding the event to the simulation, raising a warning/error if the shock on a sector is too great.
- I have added a quick check to avoid `DeprecationWarning: The truth value of an empty array is ambiguous. Returning False, but in future this will result in an error. Use `array.size > 0` to check that an array is not empty.` in the tests.
- I have also added a more direct link to pymrio's documentation on how to create custom MRIOT.

Does that correctly answer all your comments/remarks @potterzot @mwt ?
