### 2024-06-03 25d65db
misc: :fix: lexico_reindex now accounts for case when attribute is unset
