### 2024-06-03 b693126
model,simulation: improves/clean up code from testing
    - removes redundant checks
    - removed regional_production_share (unused)
    - init of several attribute with None instead of 0s
    - removed weird final_demand_cat init
    - update_prod_delta now called when changes on either sources of prod delta
    - added check on _n_rebuilding_events before setting rebuild_demand
    - several index fixes
    - simplified calc_production to avoid unnecessary/unattained case
