+++ boario/model_base.py
@@ -213,23 +213,9 @@ class ARIOBaseModel:
         #################################################################
 
         # ####### INITIAL MRIO STATE (in step temporality) ###############
-        # We could factor this with getattr but then the linter complains
-        if source_mriot.Z is None:
-            raise ValueError(
-                "Z attribute of given MRIOT doesn't exist, this is a problem"
-            )
-        if source_mriot.Y is None:
-            raise ValueError(
-                "Y attribute of given MRIOT doesn't exist, this is a problem"
-            )
-        if source_mriot.x is None:
-            raise ValueError(
-                "x attribute of given MRIOT doesn't exist, this is a problem"
-            )
-
         self._matrix_id = np.eye(self.n_sectors)
         self._matrix_I_sum = np.tile(self._matrix_id, self.n_regions)
-        self.Z_0 = source_mriot.Z.to_numpy()
+        self.Z_0 = source_mriot.Z.to_numpy() # type: ignore
         r"""numpy.ndarray of float: 2-dim square matrix array :math:`\ioz` of size :math:`(n \times m, n \times m)` representing the daily intermediate (transaction) matrix (see :ref:`boario-math-init`)."""
 
         self.Z_C = self._matrix_I_sum @ self.Z_0
@@ -294,17 +280,19 @@ class ARIOBaseModel:
             kratio_ordered = [kratio[k] for k in sorted(kratio.keys())]
             self.capital_to_VA_ratio = np.tile(np.array(kratio_ordered), self.n_regions)
             self.productive_capital = self.VA_0 * self.capital_to_VA_ratio
-        if value_added.ndim > 1:
-            self.regional_production_share = (
-                self.VA_0
-                / value_added.sum(axis=0).groupby("region").transform("sum").to_numpy()
-            )
-        else:
-            self.regional_production_share = (
-                self.VA_0 / value_added.groupby("region").transform("sum").to_numpy()
-            )
-        self.regional_production_share = self.regional_production_share.flatten()
-        r"""numpy.ndarray of float: Array of size :math:`n \times m` representing the estimated share of a sector in its regional economy."""
+
+        # Currently not used, and dubious definition
+        # if value_added.ndim > 1:
+        #     self.regional_production_share = (
+        #         self.VA_0
+        #         / value_added.sum(axis=0).groupby("region").transform("sum").to_numpy()
+        #     )
+        # else:
+        #     self.regional_production_share = (
+        #         self.VA_0 / value_added.groupby("region").transform("sum").to_numpy()
+        #     )
+        # self.regional_production_share = self.regional_production_share.flatten()
+        # r"""numpy.ndarray of float: Array of size :math:`n \times m` representing the estimated share of a sector in its regional economy."""
 
         self.threshold_not_input = (
             self.Z_C > np.tile(self.X_0, (self.n_sectors, 1)) * TECHNOLOGY_THRESHOLD
@@ -342,10 +330,11 @@ class ARIOBaseModel:
         self.final_demand = self.Y_0.copy()
         self.rebuilding_demand = None
         self._rebuild_demand_tot = np.zeros_like(self.X_0)
-        self._prod_cap_delta_arbitrary = np.zeros_like(self.X_0)
-        self._prod_cap_delta_tot = np.zeros_like(self.X_0)
+        self._prod_cap_delta_arbitrary = None # Required to init productive_capital_lost
         self.productive_capital_lost = None
         r"""numpy.ndarray of float: Array of size :math:`n \times m` representing the estimated stock of capital currently destroyed for each industry."""
+        self.prod_cap_delta_arbitrary = None
+        r"""numpy.ndarray of float: Array of size :math:`n \times m` representing an arbitrary reduction of production capacity to each industry."""
 
[…]
