### 2024-06-03 318bf4b
tests: implements new range of tests WIP
