+++ tests/test_events.py
@@ -1,14 +1,15 @@
-from boario.utils.recovery_functions import concave_recovery, convexe_recovery, convexe_recovery_scaled, linear_recovery
+from boario.utils.recovery_functions import (
+    concave_recovery,
+    convexe_recovery,
+    convexe_recovery_scaled,
+    linear_recovery,
+)
 import pytest
 
 from contextlib import nullcontext as does_not_raise
 
 # import pandas for the plot
 import pandas as pd
-import numpy as np
-
-from numpy.core._exceptions import UFuncTypeError
-
 # import the different classes
 import boario.event as event
 
@@ -139,39 +140,48 @@ class TestEventInitSeries:
 
     def test_event_from_series_wrongtype(self, sample_series):
         with pytest.raises(ValueError):
-            event.from_series(sample_series, event_type="wrongtype")
+            event.from_series(sample_series, event_type="wrongtype")  # type: ignore
 
     def test_event_from_series_notype(self, sample_series):
         with pytest.raises(TypeError):
-            event.from_series(sample_series)
+            event.from_series(sample_series)  # type: ignore
 
 
 class TestEventInitScalar:
 
     def test_event_from_scalar_arbitrary(self):
         with pytest.raises(NotImplementedError):
-            event.from_scalar_industries(impact = 1000, affected_industries=pd.Index([("RegionA","Sector1")]), event_type="arbitrary")
+            event.from_scalar_industries(impact=1000, affected_industries=pd.Index([("RegionA", "Sector1")]), event_type="arbitrary")  # type: ignore
 
     def test_event_from_scalar_wrongtype(self):
         with pytest.raises(ValueError):
-            event.from_scalar_industries(impact = 1000, affected_industries=pd.Index([("RegionA","Sector1")]), event_type="wrongtype")
+            event.from_scalar_industries(impact=1000, affected_industries=pd.Index([("RegionA", "Sector1")]), event_type="wrongtype")  # type: ignore
 
     def test_event_from_scalar_notype(self):
         with pytest.raises(TypeError):
-            event.from_scalar_industries(sample_series, affected_industries=pd.Index([("RegionA","Sector1")]))
+            event.from_scalar_industries(sample_series, affected_industries=pd.Index([("RegionA", "Sector1")]))  # type: ignore
 
     def test_event_from_scalar_reg_sec_arbitrary(self):
         with pytest.raises(NotImplementedError):
-            event.from_scalar_regions_sectors(impact = 1000, affected_regions=pd.Index(["RegionA"]), affected_sectors=pd.Index(["SectorA"]), event_type="arbitrary")
+            event.from_scalar_regions_sectors(
+                impact=1000,
+                affected_regions=pd.Index(["RegionA"]),
+                affected_sectors=pd.Index(["SectorA"]),
+                event_type="arbitrary",
+            )
 
     def test_event_from_scalar_reg_sec_wrongtype(self):
         with pytest.raises(ValueError):
-            event.from_scalar_regions_sectors(impact = 1000, affected_regions=pd.Index(["RegionA"]), affected_sectors=pd.Index(["SectorA"]), event_type="wrongtype")
+            event.from_scalar_regions_sectors(
+                impact=1000,
+                affected_regions=pd.Index(["RegionA"]),
+                affected_sectors=pd.Index(["SectorA"]),
+                event_type="wrongtype",
+            )
 
     def test_event_from_scalar_reg_sec_notype(self):
         with pytest.raises(TypeError):
-            event.from_scalar_regions_sectors(sample_series, affected_regions=pd.Index(["RegionA"]), affected_sectors=pd.Index(["SectorA"]))
-
+            event.from_scalar_regions_sectors(sample_series, affected_regions=pd.Index(["RegionA"]), affected_sectors=pd.Index(["SectorA"]))  # type: ignore
 
[…]
