+++ tests/test_models.py
@@ -1,14 +1,15 @@
+import re
 from boario.extended_models import ARIOPsiModel
 
+from boario.utils.misc import lexico_reindex
 import pytest
 
-# import pymrio for the test MRIO
+
 import pymrio
 
-# import pandas for the plot
 import numpy as np
+import pandas as pd
 
-# import the different classes
 from boario.utils.recovery_functions import *
 
 
@@ -60,19 +61,16 @@ class TestARIOPsiModel:
         assert model.had_shortage == False
 
         assert isinstance(model.rebuild_tau, (int, np.integer))
-
+        test_mrio = lexico_reindex(test_mrio)
         np.testing.assert_allclose(model.Z_0, test_mrio.Z.to_numpy() * 1 / 365)
         assert np.allclose(model.Y_0, test_mrio.Y.to_numpy() * 1 / 365)
         assert np.allclose(model.X_0, test_mrio.x.T.to_numpy().flatten() * 1 / 365)
-        assert np.allclose(model.matrix_orders, test_mrio.Z.to_numpy() * 1 / 365)
+        assert np.allclose(model.intermediate_demand, test_mrio.Z.to_numpy() * 1 / 365)
         assert np.allclose(
             model.production, test_mrio.x.T.to_numpy().flatten() * 1 / 365
         )
-        assert np.allclose(model.intmd_demand, test_mrio.Z.to_numpy() * 1 / 365)
         assert np.allclose(model.final_demand, test_mrio.Y.to_numpy() * 1 / 365)
-        assert np.allclose(
-            model.productive_capital_lost, np.zeros(shape=len(test_mrio.x.squeeze()))
-        )
+        assert model.productive_capital_lost is None
 
     # Tests creating an instance of the class with the `psi_param` parameter as a string with a valid float value.
     def test_init_psi_param_string(self, test_mrio):
@@ -95,3 +93,371 @@ class TestARIOPsiModel:
         model = ARIOPsiModel(test_mrio, psi_param=1)
         assert isinstance(model, ARIOPsiModel)
         assert model.psi == 1.0
+
+    def test_monetary_factor(self, test_mrio):
+        model = ARIOPsiModel(test_mrio, monetary_factor=10**7)
+        assert model.monetary_factor == 10**7
+
+    def test_warnings(self, test_mrio):
+        del test_mrio.meta
+        test_mrio.monetary_factor = 1
+        with pytest.warns() as record:
+            ARIOPsiModel(test_mrio, psi_param=1, iotable_year_to_temporal_unit_factor=7)
+
+        assert len(record) == 4
+        assert str(record[0].message) == "It seems the MRIOT you loaded doesn't have metadata to print."
+        assert str(record[1].message) == "Custom monetary factor found in the mrio pickle file, continuing with this one (1)"
+        assert str(record[2].message) == "iotable_to_daily_step_factor is not set to 365 (days). This should probably not be the case if the IO tables you use are on a yearly basis."
+        assert str(record[3].message) == "No capital to VA dictionary given, considering 4/1 ratio"#    "It seems the MRIOT you loaded doesn't have metadata to print."
+
+    def test_incomplete_mriot(self, test_mrio):
+        Z, Y = test_mrio.Z.copy(),test_mrio.Y.copy()
+        del test_mrio.Z
+        with pytest.raises(ValueError):
+            ARIOPsiModel(test_mrio)
+        test_mrio.Z = Z
+        del test_mrio.Y
+        with pytest.raises(ValueError):
+            ARIOPsiModel(test_mrio)
+        test_mrio.Y = Y
+        del test_mrio.x
+        with pytest.raises(ValueError):
+            ARIOPsiModel(test_mrio)
+
[…]
