### 2024-06-03 d8c64c4
model: fixes casting to float pblm with productive_capital calc
