### 2024-06-06 #6547 @crvernon
👋  @spjuhel - please make the following corrections in your paper.  After you do so, I will take another review pass.

- [ ] LINES 3 and 4:  Since you are the sole author, you do not need the designation of equal contribution
- [ ] LINE 16:  "and sometime" should be "and sometimes"
- [ ] LINE 22: "and pivotal model" should be "and a pivotal model"
- [ ] LINE 28:  missing space between "region," and "sector" in "(region,sector)"
- [ ] LINE 33-34:  "The model can then describe how the economic, as depicted, responds to a shock" where "economic" should be "economy"
- [ ] LINE 40: "only a few implementation" should be "only a few implementations"
- [ ] LINES 42-50:  No need to state the name of the person who owns the repository.  If there is a citation, just use it and not the links.
- [ ] LINE 42: "python" should have a capital "P"
- [ ] LINE 44: "python" should have a capital "P"
- [ ] LINE 51: "implementation offers" should be "implementations offer"
- [ ] LINES 33 and 60:  you define "Multi-Regional Input-Output Tables (MRIOTs)"  on line 66 but mention it first on line 33
- [ ] LINE 62: "This aspect notably allows to fully benefit from" should be "This aspect notably permits full benefit from"
- [ ] LINE 63: "such tables are becoming available" should be "such tables as the become available"
- [ ] LINE 67: "types of case-study can" should be "types of case studies can"
- [ ] LINE 69: "the recovery of from the impact" should be "the recovery from the impact"
- [ ] LINE 69:  in "from the impact, etc. but also from" you can just remove the etc and state "from the impact, and also from"
- [ ] LINE 70: "The rationale for detailed" should be "The rationale for the detailed"
- [ ] LINE 74: "DataFrames" should be "DataFrame"
- [ ] LINE 86: "inter-operable" should be "interoperable"
- [ ] LINE 89-94:  I would remove these lines completely.  I think you already make a good case for this upfront and these lines are not needed.
- [ ] LINE 100: "BoARIO can be installed from pip using:" should be "BoARIO can be installed from PyPi using:"
- [ ] LINE 102-104:  You can remove these lines altogether
- [ ] LINE 109: "and Alessio Ciullo, for its interest in the package, its valuable suggestions" should be "and Alessio Ciullo for their interest and valuable suggestions for the package"
- [ ] Throughout:  make the spelling of "modeling" or "modelling" consistent throughout
