### 2024-06-06 ecd79c7
paper.md: integrates C. Vernon remarks
