+++ paper.md
@@ -8,7 +8,6 @@ tags:
 authors:
   - name: Samuel Juhel
     orcid: 0000-0001-8801-3890
-    equal-contrib: true
     affiliation: "1, 2" # (Multiple affiliations must be quoted)
 affiliations:
  - name: CIRED, France
@@ -35,7 +34,7 @@ between multiple regions and sectors. Recent research in the field argues in
 favor of using more Agent-Based oriented model, associated with an increase in
 the complexity of the mechanisms represented [@coronese-2022-econom-impac].
 However, the assumptions and hypotheses underlying these economic mechanisms
-vary a lot, and sometime lack transparency, making it difficult to properly
+vary a lot, and sometimes lack transparency, making it difficult to properly
 interpret and compare results across models, even more so when the code used is
 not published or undocumented.
 
@@ -43,7 +42,7 @@ The Adaptive Regional Input-Output model (or ARIO) is an hybrid input-output /
 agent-based economic model, designed to compute indirect costs consequent to
 economic shocks. Its first version dates back to 2008 and was originally
 developed to assess the indirect costs of natural disasters
-[@hallegatte-2008-adapt-region]. ARIO is now a well-established and pivotal
+[@hallegatte-2008-adapt-region]. ARIO is now a well-established and a pivotal
 model in its field, has been used in multiple studies, and has seen several
 extensions or adaptations [@wu-2011-region-indir; @ranger-2010-asses-poten;
 @henriet-2012-firm-networ; @hallegatte-2013-model-role;
@@ -52,14 +51,14 @@ extensions or adaptations [@wu-2011-region-indir; @ranger-2010-asses-poten;
 @wang-2020-econom-footp; @wang-2018-quant-spatial].
 
 In ARIO, the economy is modelled as a set of economic sectors and regions, and
-we call a specific (region,sector) couple an *industry*. Each industry produces
+we call a specific (region, sector) couple an *industry*. Each industry produces
 a unique product which is assumed to be the same for all industries of the same
 sector. Each industry keeps an inventory of inputs it requires for production.
 Each industry answers a total demand consisting of the final demand (from
 households, public spendings and private investments) and of the intermediate
 demand (from other industries). An initial equilibrium state for the economy is
 built based on a multi-regional input-output table. The model can then describe
-how the economic, as depicted, responds to a shock (or multiple ones).
+how the economy, as depicted, responds to a shock (or multiple ones).
 
 `BoARIO` is an open-source Python package implementing the ARIO model. Its core
 purpose is to help support better accessibility, transparency, replicability and
@@ -68,16 +67,16 @@ comparability in the field of indirect economic impacts modeling.
 # Statement of need
 
 Although the ARIO model has been used in multiple studies, and several extensions
-exists, only a few implementation of the model or similar ones are openly available.
+exists, only a few implementations of the model or similar ones are openly available.
 We found the following existing implementations:
 
-   - A python implementation of MRIA [@koks-2016-multir-impac], is available on the [personal GitHub repository of E. Koks](https://github.com/ElcoK/MRIA).
-   - [C. Colon repository](https://github.com/ccolon/disrupt-supply-chain-model/) holds a python implementation of Disrupt Supply Chain [@colon-2020-critic-analy].
-   - A C++ implementation of the Acclimate model [@otto-2017-model-loss], is available [here](https://github.com/acclimate/acclimate).
-   - A Matlab implementation of C. Shughrue's model [@shughrue-2020-global-spread], is available on [his repository](https://github.com/chrisshughrue/GlobalUrbanCycloneImpactSimulation).
-   - The ARIO models version used in [@wang-2020-econom-footp, @guan-2020-global-suppl] are both available on [D. Wang personal repository](https://github.com/DaopingW/)
[…]
