### 2024-06-08 2e90e58
models_tests: WIP additional tests
