### 2024-06-08 #6547 @spjuhel
👋 @crvernon,

I noticed I forgot to merge my develop branch into main, hence I think you reviewed a paper version slightly not up to date. Apologies for that. (Changes are minimal, though).

I should have addressed all your comments, and I did the merge on main, so the paper is now up-to-date.
