### 2024-06-10 #6547 @crvernon
Thanks @spjuhel.  Please correct the following errors in the paper:

- [ ] LINE 92:  "french" should be "French"
- [ ] LINE 108: something is not correct with this citation (e.g., nil)
- [ ] LINE 112: "covid-19" should be "COVID-19", capitalization can be maintained by placing curly brackets around the letters or words you want to keep capitalized.
- [ ] LINE 115: "katrina" should be "Katrina"
- [ ] LINE 122: "copenhagen" should be "Copenhagen"
- [ ] LINE 149:  something is not correct with this citation (e.g., nil)
- [ ] LINE 153: "mumbai" should be "Mumbai"
- [ ] LINE 173: something is not correct with this citation (e.g., nil)
- [ ] LINE 175: "bohai sea" should be "Bohai Sea" and "china" should be "China"
- [ ] LINE 179: "california" should be "California"
- [ ] LINE 182: "wenchuan" should be "Wenchuan"

Please make sure all of the references are formatted correctly and no "nil" values exists. Thank you!
