### 2024-06-10 426c163
Small correction in paper.md
