+++ paper.bib
@@ -36,11 +36,9 @@
   Title = {Economic Impacts of Natural Hazards and Complexity Science: a Critical Review},
   Author = {Matteo Coronese and Davide Luzzati},
   DOI = {10.2139/ssrn.4101276},
-  Journal = {SSRN Electronic Journal},
-  Number = {nil},
-  Pages = {nil},
+  Journal = {LEM Working Paper},
+  number = {2022/13}
   URL = {http://dx.doi.org/10.2139/ssrn.4101276},
-  Volume = {nil},
   Year = {2022},
   date_added = {Thu Nov 23 17:17:22 2023},
 }
@@ -109,7 +107,6 @@ https://github.com/IndEcol/pymrio
   DOI = {10.1029/2020ef001616},
   JournalTitle = {Earth's Future},
   Number = {12},
-  Pages = {nil},
   URL = {https://doi.org/10.1029/2020ef001616},
   Volume = {8},
   Year = {2020},
@@ -157,7 +154,7 @@ https://github.com/IndEcol/pymrio
 }
 
 @article{guan-2020-global-suppl,
-  Title = {Global Supply-Chain Effects of Covid-19 Control Measures},
+  Title = {Global Supply-Chain Effects of {COVID}-19 Control Measures},
   Author = { Dabo Guan and Daoping Wang and Stephane Hallegatte and Steven J. Davis and Jingwen Huo and Shuping Li and Yangchun Bai and Tianyang Lei and Qianyu Xue and D'Maris Coffman and Danyang Cheng and Peipei Chen and Xi Liang and Bing Xu and Xiaosheng Lu and Shouyang Wang and Klaus Hubacek and Peng Gong },
   DOI = {10.1038/s41562-020-0896-8},
   JournalTitle = {Nature Human Behaviour},
@@ -170,7 +167,7 @@ https://github.com/IndEcol/pymrio
 }
 
 @article{ranger-2010-asses-poten,
-  Title = {An Assessment of the Potential Impact of Climate Change on Flood Risk in Mumbai},
+  Title = {An Assessment of the Potential Impact of Climate Change on Flood Risk in {M}umbai},
   Author = { Nicola Ranger and St{\'e}phane Hallegatte and Sumana Bhattacharya and Murthy Bachu and Satya Priya and K. Dhore and Farhat Rafique and P. Mathur and Nicolas Naville and Fanny Henriet and Celine Herweijer and Sanjib Pohit and Jan Corfee-Morlot },
   DOI = {10.1007/s10584-010-9979-2},
   JournalTitle = {Climatic Change},
@@ -196,7 +193,7 @@ https://github.com/IndEcol/pymrio
 }
 
 @article{hallegatte-2010-asses-climat,
-  Title = {Assessing Climate Change Impacts, Sea Level Rise and Storm Surge Risk in Port Cities: a Case Study on Copenhagen},
+  Title = {Assessing Climate Change Impacts, Sea Level Rise and Storm Surge Risk in Port Cities: a Case Study on {C}openhagen},
   Author = {St{\'e}phane Hallegatte and Nicola Ranger and Olivier Mestre and Patrice Dumas and Jan Corfee-Morlot and Celine Herweijer and Robert Muir Wood},
   DOI = {10.1007/s10584-010-9978-3},
   JournalTitle = {Climatic Change},
@@ -209,7 +206,7 @@ https://github.com/IndEcol/pymrio
 }
 
 @article{hallegatte-2008-adapt-region,
-  Title = {An Adaptive Regional Input-Output Model and Its Application To the Assessment of the Economic Cost of Katrina},
[…]
