+++ paper.bib
@@ -37,7 +37,7 @@
   Author = {Matteo Coronese and Davide Luzzati},
   DOI = {10.2139/ssrn.4101276},
   Journal = {LEM Working Paper},
-  number = {2022/13}
+  number = {2022/13},
   URL = {http://dx.doi.org/10.2139/ssrn.4101276},
   Year = {2022},
   date_added = {Thu Nov 23 17:17:22 2023},
@@ -137,13 +137,11 @@ https://github.com/IndEcol/pymrio
                   Nicolas Colombi and
                   wjan262 and
                   Simona Meiler and
-                  leonie-villiger and
-                  climada-jenkins and
+                  Leonie Villiger and
                   Rachel\_B and
-                  raphael-portmann and
-                  veronicabozzini and
-                  DarioStocker and
-                  scem},
+                  Raphael Portmann and
+                  Veronica Bozzini and
+                  Dario Stocker},
   title        = {CLIMADA-project/climada\_python: v4.0.1},
   month        = sep,
   year         = 2023,
