### 2024-06-11 #6547 @spjuhel
I think this should be good: https://doi.org/10.5281/zenodo.11580697
