### 2024-06-11 #6547 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
